### 2025-02-20 #7816 @editorialbot
[PRE REVIEW]: NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality
**Submitting author:** @hjstroud (<a href="http://orcid.org/0000-0001-8440-2629">Harry J. Stroud</a>)
**Repository:** https://github.com/IRL2/nanover-server-py
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.2788
