### 2025-02-21 c99e12c
Update bib file path in paper.md
