### 2025-03-03 853c973
Fix .NET/C# worfklow (#386)
* remove old ssl hotfix from c# action
* change ubuntu version
