### 2025-03-17 #7816 @Kevin-Mattheus-Moerman
- [ ] Please study the above reference check :point_up: and see if you can address any of the reported potential DOI issues. Furthermore, can you check if the .bib file can be purged such that it only contains works cited here? You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
