### 2025-03-18 a55b11f
Create OpenMMSMDSimulation class with very basic functionality
