+++ CONTRIBUTING.md
@@ -1,0 +1,5 @@
+# How to contribute to NanoVer
+
+The easiest way to contribute to NanoVer is to use the software and report here on GitHub and bugs and issues you run into. For code contributions it is best that we discuss first the motivations in an issue or directly over email or some other contact before you embark on any concrete work.
+
+When reporting issues please make sure you are using the latest versions of the software and include the package version if you are using it via Conda. The more specific you are about the conditions leading to the issue, the easier it is for us to determine the cause.
