### 2025-03-19 2d11052
Rework implementation of OpenMMSMDSimulation to return correct subclass upon creation using from_simulation or from_xml_path classmethods.
