+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from typing import Optional, Union, Any
 
 import numpy as np
+from ase.transport.tools import get_subspace
 
 from openmm import CustomExternalForce, CustomCentroidBondForce
 from openmm.app import Simulation
@@ -19,11 +20,20 @@ class OpenMMSMDSimulation:
         """
         Construct the SMD simulation from an existing OpenMM simulation.
         """
-        sim = cls(name)
+
+        # Create instance of SMD simulation based on type of indices passed
+        assert(smd_atom_indices.size >= 1)
+        if smd_atom_indices.size > 1:
+            sim = super(cls, OpenMMSMDSimulationCOM).__new__(OpenMMSMDSimulationCOM)
+        else:
+            sim = super(cls, OpenMMSMDSimulationAtom).__new__(OpenMMSMDSimulationAtom)
+
+        sim.name = name
         sim.simulation = simulation
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
+        sim.n_smd_atom_indices = sim.smd_atom_indices.size
 
         # Create SMD force and add it to the system
         sim.add_smd_force_to_system()
@@ -34,23 +44,33 @@ class OpenMMSMDSimulation:
 
         return sim
 
+
     @classmethod
     def from_xml_path(cls, path: PathLike[str], smd_atom_indices: np.ndarray, smd_path: np.ndarray, smd_force_constant: float, *, name: Optional[str] = None):
         """
         Construct the SMD simulation from an existing NanoVer OpenMM XML file located at a given path.
         """
+
+        if smd_atom_indices.size > 1:
+            sim = super(cls, OpenMMSMDSimulationCOM).__new__(OpenMMSMDSimulationCOM)
+        elif smd_atom_indices.size == 1:
+            sim = super(cls, OpenMMSMDSimulationAtom).__new__(OpenMMSMDSimulationAtom)
+
+
         if name is None:
-            name = Path(path).stem
-        sim = cls(name)
+            sim.name = Path(path).stem
+        else:
+            sim.name = name
+
         sim.xml_path = path
 
         # Load the simulation from the path
         with open(sim.xml_path) as infile:
             sim.simulation = serializer.deserialize_simulation(infile)
-
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
+        sim.n_smd_atom_indices = sim.smd_atom_indices.size
 
         # Create SMD force and add it to the system
         sim.add_smd_force_to_system()
@@ -73,25 +93,46 @@ class OpenMMSMDSimulation:
         self.smd_path: Optional[np.ndarray] = None
         self.smd_force_constant: Optional[float] = None
 
+        self.n_smd_atom_indices: Optional[int] = None
+
         self.smd_force: Optional[Union[CustomExternalForce, CustomCentroidBondForce]] = None
 
         self.checkpoint: Optional[Any] = None
 
 
+
+
+class OpenMMSMDSimulationAtom(OpenMMSMDSimulation):
+
+    def __init__(self, name: Optional[str] = None):
+
+        super().__init__(name)
+
+
     def add_smd_force_to_system(self):
-        """
-        Add the appropriate SMD force to the system.
-        """
-        assert(self.simulation is not None
-               and self.smd_force_constant is not None
-               and self.smd_path is not None
-               and self.smd_atom_indices is not None
-        )
 
-        self.smd_force = define_smd_force(force_constant=self.smd_force_constant,
[…]
