### 2025-03-20 ecb5bbc
Update CONTRIBUTING.md
Co-authored-by: Harry Stroud <75842225+hjstroud@users.noreply.github.com>
