### 2025-03-24 9986792
Define functions to add SMD force and update position of SMD force
