+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -1,14 +1,15 @@
 from os import PathLike
 from pathlib import Path
 from typing import Optional, Union, Any
+from abc import abstractmethod
 
 import numpy as np
-from ase.transport.tools import get_subspace
 
 from openmm import CustomExternalForce, CustomCentroidBondForce
 from openmm.app import Simulation
 
 from nanover.openmm import serializer
+
 
 class OpenMMSMDSimulation:
     """
@@ -16,13 +17,21 @@ class OpenMMSMDSimulation:
     """
 
     @classmethod
-    def from_simulation(cls, simulation: Simulation, smd_atom_indices: np.ndarray, smd_path: np.ndarray, smd_force_constant: float, *, name: Optional[str] = None):
+    def from_simulation(
+        cls,
+        simulation: Simulation,
+        smd_atom_indices: np.ndarray,
+        smd_path: np.ndarray,
+        smd_force_constant: float,
+        *,
+        name: Optional[str] = None
+    ):
         """
         Construct the SMD simulation from an existing OpenMM simulation.
         """
 
         # Create instance of SMD simulation based on type of indices passed
-        assert(smd_atom_indices.size >= 1)
+        assert smd_atom_indices.size >= 1
         if smd_atom_indices.size > 1:
             sim = super(cls, OpenMMSMDSimulationCOM).__new__(OpenMMSMDSimulationCOM)
         else:
@@ -33,7 +42,6 @@ class OpenMMSMDSimulation:
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
-        sim.n_smd_atom_indices = sim.smd_atom_indices.size
 
         # Create SMD force and add it to the system
         sim.add_smd_force_to_system()
@@ -44,9 +52,16 @@ class OpenMMSMDSimulation:
 
         return sim
 
-
     @classmethod
-    def from_xml_path(cls, path: PathLike[str], smd_atom_indices: np.ndarray, smd_path: np.ndarray, smd_force_constant: float, *, name: Optional[str] = None):
+    def from_xml_path(
+        cls,
+        path: PathLike[str],
+        smd_atom_indices: np.ndarray,
+        smd_path: np.ndarray,
+        smd_force_constant: float,
+        *,
+        name: Optional[str] = None
+    ):
         """
         Construct the SMD simulation from an existing NanoVer OpenMM XML file located at a given path.
         """
@@ -55,7 +70,6 @@ class OpenMMSMDSimulation:
             sim = super(cls, OpenMMSMDSimulationCOM).__new__(OpenMMSMDSimulationCOM)
         elif smd_atom_indices.size == 1:
             sim = super(cls, OpenMMSMDSimulationAtom).__new__(OpenMMSMDSimulationAtom)
-
 
         if name is None:
             sim.name = Path(path).stem
@@ -70,7 +84,6 @@ class OpenMMSMDSimulation:
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
-        sim.n_smd_atom_indices = sim.smd_atom_indices.size
 
         # Create SMD force and add it to the system
         sim.add_smd_force_to_system()
@@ -81,8 +94,6 @@ class OpenMMSMDSimulation:
 
         return sim
 
-
-
     def __init__(self, name: Optional[str] = None):
         self.name = name or "Unnamed OpenMM SMD Simulation"
 
@@ -92,14 +103,23 @@ class OpenMMSMDSimulation:
         self.smd_atom_indices: Optional[np.ndarray] = None
[…]
