### 2025-03-24 30fdfe3
Add function to retrieve positions of the atoms with which the SMD interactions occur and save them to an array.
