### 2025-03-24 00c0ecc
Add necessary functions to calculate the work done on the system along the reaction coordinate (need to add tests for them)
