+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -42,8 +42,12 @@ class OpenMMSMDSimulation:
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
+        sim.n_smd_atom_indices = sim.smd_atom_indices.size
+        sim.define_smd_simulation_atom_positions_array()
 
         # Create SMD force and add it to the system
+        sim.current_smd_force_position = sim.smd_path[0]
+        sim.current_smd_force_position_index = 0
         sim.add_smd_force_to_system()
         sim.simulation.context.reinitialize(preserveState=True)
 
@@ -85,8 +89,11 @@ class OpenMMSMDSimulation:
         sim.n_smd_atom_indices = sim.smd_atom_indices.size
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
+        sim.define_smd_simulation_atom_positions_array()
 
         # Create SMD force and add it to the system
+        sim.current_smd_force_position = sim.smd_path[0]
+        sim.current_smd_force_position_index = 0
         sim.add_smd_force_to_system()
         sim.simulation.context.reinitialize(preserveState=True)
 
@@ -116,6 +123,17 @@ class OpenMMSMDSimulation:
         self.current_smd_force_position: Optional[np.ndarray] = None
         self.current_smd_force_position_index: Optional[int] = None
         self.smd_simulation_atom_positions: Optional[np.ndarray] = None
+        self.smd_simulation_forces: Optional[np.ndarray] = None
+        self.smd_simulation_work_done: Optional[np.ndarray] = None
+
+
+    @abstractmethod
+    def define_smd_simulation_atom_positions_array(self):
+        """
+        Define the array to which the positions of the atoms with which the
+        SMD force interacts will be saved over the course of the SMD simulation.
+        """
+        pass
 
 
     @abstractmethod
@@ -134,6 +152,14 @@ class OpenMMSMDSimulation:
         """
         pass
 
+    @abstractmethod
+    def calculate_cumulative_work_done(self):
+        """
+        Calculate the cumulative work done along the reaction coordinate
+        during the SMD simulation.
+        """
+        pass
+
 
     def get_smd_atom_positions(self):
         """
@@ -144,19 +170,48 @@ class OpenMMSMDSimulation:
         self.smd_simulation_atom_positions[self.current_smd_force_position_index] = positions[self.smd_atom_indices]
 
 
+    def run_smd(self):
+        """
+        Perform an SMD simulation on the system using the SMD force and
+        path defined, store the positions of the atoms with which the
+        SMD force interacts, and calculate the work done along the reaction
+        coordinate.
+        """
+        return None
+
+
+    def calculate_forces(self, interaction_centre_positions):
+        """
+        Calculate the SMD forces that acted on the system during the simulation.
+        """
+        assert(np.all(self.smd_path.shape == interaction_centre_positions.shape))
+        self.smd_simulation_forces = - self.smd_force_constant * (interaction_centre_positions - self.smd_path)
+
+
+    def _calculate_work_done(self):
+        """
+        Calculate the cumulative work done along the reaction coordinate
+        """
+        assert(self.smd_path is not None and self.smd_simulation_forces is not None)
+        smd_force_displacements = np.diff(self.smd_path, axis=0)
+        work_done_array = np.zeros(smd_force_displacements.shape[0])
+        for i in range(work_done_array.size):
[…]
