### 2025-03-25 c2c6302
Add function to perform SMD simulation
