+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -24,7 +24,7 @@ class OpenMMSMDSimulation:
         smd_path: np.ndarray,
         smd_force_constant: float,
         *,
-        name: Optional[str] = None
+        name: Optional[str] = None,
     ):
         """
         Construct the SMD simulation from an existing OpenMM simulation.
@@ -64,7 +64,7 @@ class OpenMMSMDSimulation:
         smd_path: np.ndarray,
         smd_force_constant: float,
         *,
-        name: Optional[str] = None
+        name: Optional[str] = None,
     ):
         """
         Construct the SMD simulation from an existing NanoVer OpenMM XML file located at a given path.
@@ -126,7 +126,6 @@ class OpenMMSMDSimulation:
         self.smd_simulation_forces: Optional[np.ndarray] = None
         self.smd_simulation_work_done: Optional[np.ndarray] = None
 
-
     @abstractmethod
     def define_smd_simulation_atom_positions_array(self):
         """
@@ -135,7 +134,6 @@ class OpenMMSMDSimulation:
         """
         pass
 
-
     @abstractmethod
     def add_smd_force_to_system(self):
         """
@@ -143,7 +141,6 @@ class OpenMMSMDSimulation:
         SMD interaction required (single atom or centre-of-mass).
         """
         pass
-
 
     @abstractmethod
     def update_smd_force_position(self):
@@ -166,40 +163,78 @@ class OpenMMSMDSimulation:
         Retrieve the positions of the atoms with which the SMD force is
         interacting, and add them to the array of positions to save.
         """
-        positions = self.simulation.context.getState(getPositions=True).getPositions(asNumpy=True)
-        self.smd_simulation_atom_positions[self.current_smd_force_position_index] = positions[self.smd_atom_indices]
+        positions = self.simulation.context.getState(getPositions=True).getPositions(
+            asNumpy=True
+        )
+        self.smd_simulation_atom_positions[self.current_smd_force_position_index] = (
+            positions[self.smd_atom_indices]
+        )
 
-
-    def run_smd(self):
+    def run_smd(self, progress_interval: Optional[int] = 100000):
         """
         Perform an SMD simulation on the system using the SMD force and
         path defined, store the positions of the atoms with which the
         SMD force interacts, and calculate the work done along the reaction
         coordinate.
         """
-        return None
 
+        # Get initial atom positions for SMD force at initial position
+        assert(self.smd_simulation_atom_positions is not None and np.all(
+            self.current_smd_force_position == self.smd_path[0]) and
+            self.current_smd_force_position_index == 0
+    )
+        self.get_smd_atom_positions()
+
+        # Run SMD procedure
+        n_steps = self.smd_path.shape[0]
+
+        print("Starting SMD simulation...")
+        for step in range(1, n_steps):
+
+            # Update force position index
+            self.current_smd_force_position_index = step
+
+            # Update SMD force position
+            self.update_smd_force_position()
+
+            # Perform single simulation step
+            self.simulation.step(1)
+
+            # Retrieve atom positions on step
+            self.get_smd_atom_positions()
+
+            # Print every 10000 steps
+            if step % progress_interval == 0:
+                print(f"Steps completed: {step}")
+
+        print("SMD simulation completed. Calculating work done...")
+
+        # Calculate the work done along the reaction coordinate
+        self.calculate_cumulative_work_done()
+
+        print("Work done calculated.")
 
     def calculate_forces(self, interaction_centre_positions):
         """
[…]
