### 2025-03-25 e750dde
Add functions to save SMD simulation data and general SMD data
