### 2025-03-25 32b33ff
Fix calculate_com_trajectory function
