### 2025-04-01 07dc413
Add zeroth element to cumulative work done
