### 2025-04-01 b0edb54
Add function to run initial equilibration of system with restraint located at initial point on SMD force path
