### 2025-04-03 4bb4595
Reset work and iMD forces when reseting simulations (#376)
* fix: reset work and imd when resetting simulation

* expand ase work tests to include resets

* expand openmm work tests to include resets, fix force resets in ImdForceManager
