### 2025-04-03 63174c2
enforcePeriodiccBox set to True
