### 2025-04-03 c2929bf
add a property for enforcing pbc in openmm sims
