### 2025-04-03 #7816 @hjstroud
@Kevin-Mattheus-Moerman sorry for the delay! I have checked over the references and they are all ok, all references in the bibfile are cited in the work. We are currently writing a contributing file, and will push it to the repo very shortly, I will send another message when this is complete.
