### 2025-04-03 #7816 @hjstroud
Also, we have further examined the list of JOSS reviewers and would like to add further suggestions for potential reviewers:

- Samuel Murail
- Jakub Rydzewski
- Alexander Rose
