### 2025-04-03 b7d8229
add a test for pbc wrapping option in OpenMMSimulation
