### 2025-04-03 4227f5c
Start functions to save simulations to xml files and generate starting structures for an SMD run
