### 2025-04-03 #7816 @hjstroud
> [@abhishektiwari](https://github.com/abhishektiwari) please wait for the authors to address the above before moving on thanks.

@Kevin-Mattheus-Moerman @ abhishektiwari We have now added a CONTRIBUTING.md file to the repo, so I think we have now addressed all of the points above. Please let us know if there's anything else we need to do!
