### 2025-04-03 fabde43
Add line in README.md linking to CONTRIBUTING.md
