### 2025-04-03 e175d04
Define function to save the simulation with the option to include the state of the system and the SMD force itself, and define an abstract function to remove the force
