### 2025-04-03 8e3d244
Define function to remove SMD forces from the system
