### 2025-04-03 1343e4c
Edit save simulation function to manually remove parameters in the case that SMD force is removed, and finish initial version of generate_starting_structures function
