+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -165,14 +165,16 @@ class OpenMMSMDSimulation:
         forces = self.simulation.system.getForces()
         forces_to_remove = []
         for i in range(len(forces)):
-            if type(forces[i]) == type(self.smd_force) and forces[i].getGlobalParameterName(0)=="smd_k":
+            if (
+                type(forces[i]) == type(self.smd_force)
+                and forces[i].getGlobalParameterName(0) == "smd_k"
+            ):
                 forces_to_remove.append(i)
 
         # Remove any SMD forces, accounting for the changes in indices
         # as forces are removed
         for j in range(len(forces_to_remove)):
-            self.simulation.system.removeForce(forces_to_remove[j]-j)
-
+            self.simulation.system.removeForce(forces_to_remove[j] - j)
 
     def get_smd_atom_positions(self):
         """
@@ -196,27 +198,37 @@ class OpenMMSMDSimulation:
         try:
             assert np.all(self.current_smd_force_position == self.smd_path[0])
         except AssertionError:
-            raise AssertionError("Restraint is not located at the initial position "
-                                 "of the SMD force, equilibration aborted. To prepare the system "
-                                 "appropriately, the restraint should be placed at the first point "
-                                 "on the path that the SMD force will take during the SMD simulation.")
+            raise AssertionError(
+                "Restraint is not located at the initial position "
+                "of the SMD force, equilibration aborted. To prepare the system "
+                "appropriately, the restraint should be placed at the first point "
+                "on the path that the SMD force will take during the SMD simulation."
+            )
 
         self.simulation.step(n_steps)
 
-
-    def save_simulation(self, output_filepath: PathLike[str], save_state: bool = False, save_smd_force: Optional[bool] = False):
+    def save_simulation(
+        self,
+        output_filepath: PathLike[str],
+        save_state: bool = False,
+        save_smd_force: Optional[bool] = False,
+    ):
         """
-        Save the simulation to a NanoVer XML file, with the option to include the
+        Save the simulation to a NanoVer OpenMM XML file, with the option to include the
         SMD force in the XML file.
         :param output_filepath: Path to output file to save the simulation to.
         :param save_state: If True, save the present state of the simulation to the XML file.
-        :param save_smd_force: Bool defining whether to save the SMD force in the XML file (optional).
+        :param save_smd_force: Bool defining whether to save the SMD force in the XML file (Optional).
         """
         assert output_filepath is not None
 
         if save_smd_force:
             with open(output_filepath, "w") as outfile:
-                outfile.write(serializer.serialize_simulation(self.simulation, save_state=save_state))
+                outfile.write(
+                    serializer.serialize_simulation(
+                        self.simulation, save_state=save_state
+                    )
+                )
 
         else:
             # Temporarily remove SMD force from simulation
@@ -224,17 +236,28 @@ class OpenMMSMDSimulation:
 
             # Save simulation without SMD force
             with open(output_filepath, "w") as outfile:
-                xml_string = serializer.serialize_simulation(self.simulation, save_state=save_state)
+                xml_string = serializer.serialize_simulation(
+                    self.simulation, save_state=save_state
+                )
                 # Manually remove parameters for now...
-                # TODO: Work out how to deal with these unwanted parameters... maybe change to per bond parameter?
-                xml_string =  "\n".join(x for x in xml_string.splitlines() if "smd_k" not in x)
+                # TODO: Work out how to deal with these unwanted parameters...
+                #  WARNING: this will not work if you have other parameters!
+                #  maybe change smd_k to a per bond parameter?
[…]
