### 2025-04-04 6f35bc7
Add more documentation for the SMD class and subclasses
