### 2025-04-04 7c083e8
Update wording of docstrings
