### 2025-04-05 #7816 @abhishektiwari
@hjstroud Thanks for adding the contribution guide and providing suggestion for potential reviewers.

Two cited articles _Molecular dynamics by electronic computers_ (published in 1957) and _Markov chain Monte Carlo maximum likelihood_ do not have DOI.

@Kevin-Mattheus-Moerman I think we can move forward to next step.
