### 2025-04-07 1dbd023
move PBC change line to be next to sim initialisation for clarity
