### 2025-04-07 67daadc
Start constructing PathSmoother class
