+++ python-libraries/nanover-server/src/nanover/smd/pathsmoothing.py
@@ -1,0 +1,373 @@
+import numpy as np
+import matplotlib.pyplot as plt
+import MDAnalysis as mda
+import ipywidgets as widgets
+
+from typing import Optional, Union, Tuple
+from os import PathLike
+from scipy.interpolate import splprep, splev
+from ipywidgets import interact, interactive, fixed, interact_manual
+
+from nanover.mdanalysis import NanoverParser, NanoverReader
+
+
+class PathSmoother:
+    """
+    A class for smoothing trajectories of atoms or groups of atoms defined
+    by user interactions during an iMD simulation.
+    """
+
+    @classmethod
+    def from_traj_file(cls, filename: Union[str, PathLike]):
+        """
+        Create a PathSmoother object from the data stored in a NanoVer trajectory file.
+
+        :param filename: Path to the trajectory file.
+        """
+        ps = cls()
+        ps.filename = filename
+        ps.retrieve_traj_data()
+        ps.atom_positions, ps.com_positions, ps.atom_indices, ps.n_interaction_frames = retrieve_uf_com_path(
+            ps.all_atom_positions, ps.all_atom_masses, ps.all_user_forces
+        )
+        return ps
+
+    def __init__(self):
+        self.filename: Optional[Union[str, PathLike]] = None
+        self.com_positions: Optional[np.ndarray] = None
+        self.atom_positions: Optional[np.ndarray] = None
+        self.atom_indices: Optional[np.ndarray] = None
+        self.n_interaction_frames: Optional[int] = None
+
+        self.all_atom_positions: Optional[np.ndarray] = None
+        self.all_user_forces: Optional[np.ndarray] = None
+        self.all_atom_masses: Optional[np.ndarray] = None
+
+        self.universe: Optional[mda.Universe] = None
+
+    def create_mda_universe(self):
+        """
+        Create an MDAnalysis universe from the given trajectory file.
+        """
+        assert self.filename is not None
+        self.universe = mda.Universe(
+            self.filename,
+            format=NanoverReader,
+            topology_format=NanoverParser,
+        )
+
+    def read_mda_universe_data(self):
+        """
+        Retrieve the relevant data for the trajectory from the MDAnalysis universe.
+        """
+        assert self.universe is not None
+
+        # Retrieve atom elements and guess masses
+        atom_elements = np.array(self.universe.atoms.types)
+        self.all_atom_masses = mda.topology.guessers.guess_masses(atom_elements)
+
+        # Create empty arrays for atom positions and user forces
+        n_atoms = atom_elements.size
+        self.all_atom_positions = np.zeros(
+            (self.universe.trajectory.n_frames - 1, n_atoms, 3)
+        )
+        self.all_user_forces = np.zeros(
+            (self.universe.trajectory.n_frames - 1, n_atoms, 3)
+        )
+
+        # Read trajectory and retrieve atom positions and user forces
+        for timestep in self.universe.trajectory:
+            # Exclude topology frame where no user forces are applied
+            if int(timestep.frame) != 0:
+                index = int(timestep.frame) - 1
+                self.all_atom_positions[index] = self.universe.atoms.positions
+                self.all_user_forces[index] = timestep.data["user_forces"]
+
+    def retrieve_traj_data(self):
+        """
+        Retrieve data about the trajectory from the given trajectory file.
+        """
+        self.create_mda_universe()
+        self.read_mda_universe_data()
+
+
+    def plot_com_trajectory(self):
+        """
+        Plot the trajectory of the COM of the atoms defining the path.
+        """
+        assert self.com_positions is not None and self.n_interaction_frames is not None
+        plot_com_trajectory(self.com_positions, self.n_interaction_frames)
+
[…]
