### 2025-04-07 #7816 @abhishektiwari
@hjstroud Before I start search for reviewers, can you please confirm if all 61 references in `paper.bib` are cited in paper? I can see there 67 inline citations in `paper.md` but I am not sure if there is any unused reference which can be purged.
