### 2025-04-08 #7816 @hjstroud
@abhishektiwari I can confirm that all 61 references are cited in the paper.
