### 2025-04-08 3fd7568
rename to `use_pbc_wrapping`
