### 2025-04-08 aa25bb9
update ubuntu version for c# CI and hope it works
