### 2025-04-09 #7816 @abhishektiwari
👋 @richardjgowers, would you be willing to review this submission for JOSS?  The submission under consideration is "NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality".

I see you have edited and reviewed for JOSS in past so I assume you are familiar with [checklist-driven reviews](https://joss.readthedocs.io/en/latest/review_criteria.html).

Given your  experience in Molecular Dynamic space, I think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks for your help.
Abhi
