### 2025-04-09 #7816 @abhishektiwari
👋 @jakryd would you be willing to review this submission for JOSS? The submission under consideration is "NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality".

I see you have reviewed for JOSS in past so I assume you are familiar with [checklist-driven reviews](https://joss.readthedocs.io/en/latest/review_criteria.html).

If you are unavailable for review, could you recommend any potential reviewers?

Abhi
