### 2025-04-09 6fac134
feature: option to wrap openmm positions in frames based on pbc (#391)
