### 2025-04-10 e239a01
Add functionality to read an SMD force already present in a loaded simulation
