### 2025-04-16 e8e6dae
redo reading for easier recording splitting idea
