### 2025-04-19 #7816 @abhishektiwari
👋 @samuelmurail @helmutcarter @Anshuman5, would any of you be willing to review this submission for JOSS?  The submission under consideration is "NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality".

Based on your experience, we think you might be able to provide a great review of this submission. If not, could you recommend any potential reviewers?

Thanks
Abhi
