### 2025-04-21 374d169
tests: use dataclasses for simulation fixtures and dynamically compute expected force vector for `test_sparse_user_forces_elements`
