+++ python-libraries/nanover-server/tests/test_omni/common.py
@@ -1,16 +1,31 @@
 from contextlib import contextmanager
+from dataclasses import dataclass
 from pathlib import Path
+from typing import TypeVar, List
+
 from nanover.app import (
     NanoverImdClient,
     NanoverApplicationServer,
     NanoverImdApplication,
 )
+from nanover.imd import ParticleInteraction
 from nanover.omni import OmniRunner
+from nanover.omni.omni import Simulation
 
 EXAMPLES_PATH = Path(__file__).parent
 RECORDING_PATH_TRAJ = EXAMPLES_PATH / "nanotube-example-recording.traj"
 RECORDING_PATH_STATE = EXAMPLES_PATH / "nanotube-example-recording.state"
 ARGON_XML_PATH = EXAMPLES_PATH / "argon_simulation.xml"
+
+
+SimulationType = TypeVar("SimulationType", bound=Simulation)
+
+
+@dataclass(kw_only=True)
+class SimulationTestData:
+    simulation: SimulationType
+    app_server: NanoverApplicationServer
+    interactions: List[ParticleInteraction]
 
 
 @contextmanager
@@ -22,7 +37,7 @@ def make_loaded_sim(sim):
 
 
 @contextmanager
-def make_loaded_sim_with_interactions(sim, *interactions):
+def make_loaded_sim_with_interactions(sim, *interactions: ParticleInteraction):
     with make_app_server() as app_server:
         sim.load()
         sim.reset(app_server)
@@ -30,7 +45,11 @@ def make_loaded_sim_with_interactions(sim, *interactions):
         for i, interaction in enumerate(interactions):
             app_server.imd.insert_interaction(f"interaction.{i}", interaction)
 
-        yield app_server, sim
+        yield SimulationTestData(
+            simulation=sim,
+            app_server=app_server,
+            interactions=list(interactions),
+        )
 
 
 @contextmanager
