+++ python-libraries/nanover-server/tests/test_omni/test_openmm.py
@@ -43,8 +43,8 @@ def make_example_openmm():
 
 @pytest.fixture
 def single_atom_app_and_simulation_with_constant_force():
-    with make_single_atom_app_and_simulation_with_constant_force() as (app_server, sim):
-        yield app_server, sim
+    with make_single_atom_app_and_simulation_with_constant_force() as data:
+        yield data
 
 
 @contextmanager
@@ -59,16 +59,16 @@ def make_single_atom_app_and_simulation_with_constant_force():
             particles=[0],
             scale=1,
         ),
-    ) as (app_server, sim):
-        yield app_server, sim
+    ) as data:
+        yield data
 
 
 @pytest.fixture
 def basic_system_app_and_simulation():
     with make_loaded_sim_with_interactions(
         OpenMMSimulation.from_simulation(build_basic_simulation())
-    ) as (app_server, sim):
-        yield app_server, sim
+    ) as data:
+        yield data
 
 
 @pytest.fixture
@@ -90,8 +90,8 @@ def basic_system_app_and_simulation_with_constant_force():
             particles=[6],
             scale=1,
         ),
-    ) as (app_server, sim):
-        yield app_server, sim
+    ) as data:
+        yield data
 
 
 @pytest.fixture
@@ -106,8 +106,8 @@ def basic_system_app_and_simulation_with_constant_force_old():
             particles=[0, 4],
             scale=1,
         ),
-    ) as (app_server, sim):
-        yield app_server, sim
+    ) as data:
+        yield data
 
 
 @pytest.fixture
@@ -124,8 +124,8 @@ def basic_system_app_and_simulation_with_complex_interactions():
             particles=(4, 5),
             interaction_type="spring",
         ),
-    ) as (app_server, sim):
-        yield app_server, sim
+    ) as data:
+        yield data
 
 
 def test_auto_force():
@@ -188,7 +188,8 @@ def test_work_done_server(single_atom_app_and_simulation_with_constant_force):
     # an Argon atom should be 20.0 kJ mol-1 analytically. Allowing for numerical error,
     # the expected value should be slightly greater than 20.0 kJ mol-1
 
-    app, sim = single_atom_app_and_simulation_with_constant_force
+    app = single_atom_app_and_simulation_with_constant_force.app_server
+    sim = single_atom_app_and_simulation_with_constant_force.simulation
 
     for _ in range(3):
         # Add step to account for zeroth (topology) frame where force is not applied
@@ -211,7 +212,8 @@ def test_work_done_frame(basic_system_app_and_simulation_with_complex_interactio
     Test that the calculated user work done on a system that appears in the frame is equal
     to the user work done as calculated in the OpenMMSimulation.
     """
-    app, sim = basic_system_app_and_simulation_with_complex_interactions
+    app = basic_system_app_and_simulation_with_complex_interactions.app_server
+    sim = basic_system_app_and_simulation_with_complex_interactions.simulation
 
     for _ in range(3):
         sim.reset(app)
@@ -229,7 +231,7 @@ def test_save_state_basic_system(basic_system_app_and_simulation_with_constant_f
     by testing that the velocities of the simulation being serialized approximately
     equal those of the simulation loaded after serialization/deserialization.
     """
-    app, sim = basic_system_app_and_simulation_with_constant_force
+    sim = basic_system_app_and_simulation_with_constant_force.simulation
 
     # Run the simulation for a few steps
     for _ in range(11):
@@ -261,7 +263,8 @@ def test_instantaneous_temperature_no_interaction(basic_system_app_and_simulatio
     instantaneous temperature calculated by the StateDataReporter of OpenMM for a
     simulation without iMD interactions.
     """
-    app, sim = basic_system_app_and_simulation
[…]
