### 2025-04-24 bdf74e4
fix: implement and test raising InvalidInteractionError when attempting to compute iMD forces with particles indexes that are out of bounds.
