### 2025-04-28 #8118 @abhishektiwari
Done. @samuelmurail - Your review will be checklist based.  First of all you need to run this command in a separate comment to create the checklist:

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.
