### 2025-04-30 10c0587
Edit classes and SMD forces to account for whether OpenMM simulation uses PBCs
