+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -45,6 +45,9 @@ class OpenMMSMDSimulation:
 
         sim.name = name
         sim.simulation = simulation
+        # Check if simulation employs periodic boundary conditions
+        sim._sim_uses_pbcs = sim.simulation.system.usesPeriodicBoundaryConditions()
+
         sim.smd_atom_indices = smd_atom_indices
         sim.smd_path = smd_path
         sim.smd_force_constant = smd_force_constant
@@ -98,6 +101,9 @@ class OpenMMSMDSimulation:
         # Load the simulation from the path
         with open(sim.xml_path) as infile:
             sim.simulation = serializer.deserialize_simulation(infile)
+        # Check if simulation employs periodic boundary conditions
+        sim._sim_uses_pbcs = sim.simulation.system.usesPeriodicBoundaryConditions()
+
         sim.smd_atom_indices = smd_atom_indices
         sim.n_smd_atom_indices = sim.smd_atom_indices.size
         sim.smd_path = smd_path
@@ -145,6 +151,8 @@ class OpenMMSMDSimulation:
         self.smd_simulation_atom_positions: Optional[np.ndarray] = None
         self.smd_simulation_forces: Optional[np.ndarray] = None
         self.smd_simulation_work_done: Optional[np.ndarray] = None
+
+        self._sim_uses_pbcs: Optional[bool] = None
 
     @abstractmethod
     def define_smd_simulation_atom_positions_array(self):
@@ -479,7 +487,8 @@ class OpenMMSMDSimulationAtom(OpenMMSMDSimulation):
             smd_force = self.simulation.system.getForce(n_forces - 1)
             params = smd_force.getParticleParameters(0)
             assert(type(smd_force) == CustomExternalForce)
-            assert(smd_force.getEnergyFunction() == "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2")
+            assert(smd_force.getEnergyFunction() == "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2" or
+                   smd_force.getEnergyFunction() == "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)")
             assert(smd_force.getNumParticles() == 1)
             assert(force_constant == self.smd_force_constant)
             assert(params[0] == self.smd_atom_indices)
@@ -498,7 +507,7 @@ class OpenMMSMDSimulationAtom(OpenMMSMDSimulation):
         """
 
         x0, y0, z0 = self.current_smd_force_position
-        smd_force = smd_single_atom_force(self.smd_force_constant)
+        smd_force = smd_single_atom_force(self.smd_force_constant, self._sim_uses_pbcs)
         smd_force.addParticle(self.smd_atom_indices, [x0, y0, z0])
         self.smd_force = smd_force
         self.simulation.system.addForce(self.smd_force)
@@ -562,7 +571,7 @@ class OpenMMSMDSimulationCOM(OpenMMSMDSimulation):
         """
 
         x0, y0, z0 = self.current_smd_force_position
-        smd_force = smd_com_force(self.smd_force_constant)
+        smd_force = smd_com_force(self.smd_force_constant, self._sim_uses_pbcs)
         smd_force.addGroup(self.smd_atom_indices)
         smd_force.addBond([0], [x0, y0, z0])
         self.smd_force = smd_force
@@ -627,7 +636,7 @@ class OpenMMSMDSimulationCOM(OpenMMSMDSimulation):
         self._calculate_work_done()
 
 
-def smd_com_force(force_constant: float):
+def smd_com_force(force_constant: float, uses_pbcs: bool):
     """
[…]
