### 2025-04-30 f8dc18d
Edit check existing smd force function to include periodic and nonperiodic cases
