+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -487,8 +487,8 @@ class OpenMMSMDSimulationAtom(OpenMMSMDSimulation):
             smd_force = self.simulation.system.getForce(n_forces - 1)
             params = smd_force.getParticleParameters(0)
             assert(type(smd_force) == CustomExternalForce)
-            assert(smd_force.getEnergyFunction() == "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2" or
-                   smd_force.getEnergyFunction() == "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)")
+            assert(smd_force.getEnergyFunction() == "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2"
+                   or smd_force.getEnergyFunction() == "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)")
             assert(smd_force.getNumParticles() == 1)
             assert(force_constant == self.smd_force_constant)
             assert(params[0] == self.smd_atom_indices)
