### 2025-04-30 39e1f7b
Document bool to specify whether SMD force uses PBCs
