### 2025-05-02 365133d
Add classmethod to create PathSmoother class from a set of coordinates and atomic masses
