### 2025-05-02 9e0178a
Add functionality to make plots interactive in Jupyter or IPython interfaces, and functionality to close open plots
