### 2025-05-05 b5bf89b
add parameter to enforce_pbc in a .xml
