+++ python-libraries/nanover-server/tests/test_openmm/test_serializer.py
@@ -6,7 +6,7 @@ Tests for :mod:`nanover.openmm.serializer`.
 # pylint: disable=redefined-outer-name,unused-import
 from xml.dom.minidom import parseString
 import pytest
-
+from io import StringIO
 from nanover.openmm.imd import create_imd_force
 from nanover.openmm.serializer import (
     serialize_simulation,
@@ -141,16 +141,6 @@ def test_serializer_pbc():
     omm_sim = build_basic_simulation()
     UNIT_SIMULATION_BOX_VECTORS = [[1, 0, 0], [0, 1, 0], [0, 0, 1]]
     omm_sim.context.setPeriodicBoxVectors(*UNIT_SIMULATION_BOX_VECTORS)
-    sim = OpenMMSimulation.from_simulation(omm_sim)
-    sim.load()
-    with NanoverImdApplication.basic_server(port=0) as app_server:
-        sim.reset(app_server)
-        sim.advance_by_one_step()
-
-    with open("test-sim-pbc.xml", "w") as outfile:
-        outfile.write(serialize_simulation(omm_sim, pbc_wrapping=True))
-    with open("test-sim-no-pbc.xml", "w") as outfile:
-        outfile.write(serialize_simulation(omm_sim, pbc_wrapping=False))
 
     def out_of_bounds(coord):
         return coord < 0 or coord > 1
@@ -160,11 +150,13 @@ def test_serializer_pbc():
             for coord in position:
                 yield coord
 
-    with open("test-sim-pbc.xml") as infile:
-        sim_pbc_obj = deserialize_simulation(infile.read())
+    with StringIO() as xml_pbc:
+        xml_pbc.write(serialize_simulation(omm_sim, pbc_wrapping=True))
+        xml_pbc.seek(0)
+        sim_pbc_obj = deserialize_simulation(xml_pbc.read())
+
     openmm_sim = OpenMMSimulation.from_simulation(sim_pbc_obj)
     openmm_sim.load()
-
     with NanoverImdApplication.basic_server(port=0) as app_server:
         openmm_sim.reset(app_server)
 
@@ -172,11 +164,14 @@ def test_serializer_pbc():
         out_of_bounds(coord) for coord in get_sim_position_coords(openmm_sim)
     )
 
-    with open("test-sim-no-pbc.xml") as infile:
-        sim_pbc_obj_1 = deserialize_simulation(infile.read())
-    openmm_sim_1 = OpenMMSimulation.from_simulation(sim_pbc_obj_1)
-    openmm_sim_1.load()
-    with NanoverImdApplication.basic_server(port=0) as app_server:
-        openmm_sim_1.reset(app_server)
+    with StringIO() as xml_no_pbc:
+        xml_no_pbc.write(serialize_simulation(omm_sim, pbc_wrapping=False))
+        xml_no_pbc.seek(0)
+        sim_no_pbc_obj = deserialize_simulation(xml_no_pbc.read())
 
-    assert any(out_of_bounds(coord) for coord in get_sim_position_coords(openmm_sim_1))
+    openmm_sim = OpenMMSimulation.from_simulation(sim_no_pbc_obj)
+    openmm_sim.load()
+    with NanoverImdApplication.basic_server(port=0) as app_server:
+        openmm_sim.reset(app_server)
+
+    assert any(out_of_bounds(coord) for coord in get_sim_position_coords(openmm_sim))
