### 2025-05-14 4448e7d
Minor changes to pathsmoothing.py
