### 2025-05-14 1354064
Add the option to speficy whether aspect ratios of the plotted trajectories should be equal or not
