### 2025-05-15 96acf35
Add colour map options to plots and format with black
