+++ python-libraries/nanover-server/src/nanover/smd/pathsmoothing.py
@@ -14,7 +14,11 @@ from nanover.mdanalysis import NanoverParser, NanoverReader
 try:
     from ipywidgets import interact, interactive, fixed, interact_manual
     from IPython import get_ipython
-    _in_notebook = get_ipython().__class__.__name__ in ['ZMQInteractiveShell', 'TerminalInteractiveShell']
+
+    _in_notebook = get_ipython().__class__.__name__ in [
+        "ZMQInteractiveShell",
+        "TerminalInteractiveShell",
+    ]
 except:
     _in_notebook = False
 
@@ -65,7 +69,10 @@ class PathSmoother:
             ps.atom_positions = coords
         ps.atom_masses = atom_masses
         ps.com_positions = np.array(
-            [calculate_com(ps.atom_positions[i], ps.atom_masses) for i in range(ps.atom_positions.shape[0])]
+            [
+                calculate_com(ps.atom_positions[i], ps.atom_masses)
+                for i in range(ps.atom_positions.shape[0])
+            ]
         )
         ps.n_interaction_frames = ps.atom_positions.shape[0]
 
@@ -79,8 +86,10 @@ class PathSmoother:
         if _in_notebook:
             get_ipython().run_line_magic("matplotlib", "widget")
         else:
-            print("WARNING: interactive plots are only supported for Jupyter or IPython "
-                  "interfaces.")
+            print(
+                "WARNING: interactive plots are only supported for Jupyter or IPython "
+                "interfaces."
+            )
 
     @staticmethod
     def close_interactive_plots():
@@ -89,9 +98,8 @@ class PathSmoother:
         or IPython session.
         """
         if _in_notebook:
-            plt.close('all')
+            plt.close("all")
             print("Existing interactive plots closed.")
-
 
     def __init__(self):
         self.filename: Optional[Union[str, PathLike]] = None
@@ -160,7 +168,9 @@ class PathSmoother:
         :param equal_aspect_ratio: A bool defining whether the axes should have equal aspect ratio
         """
         assert self.com_positions is not None and self.n_interaction_frames is not None
-        plot_com_trajectory(self.com_positions, self.n_interaction_frames, equal_aspect_ratio)
+        plot_com_trajectory(
+            self.com_positions, self.n_interaction_frames, equal_aspect_ratio
+        )
 
     def plot_atoms_trajectories(self, equal_aspect_ratio: bool = False):
         """
@@ -169,7 +179,9 @@ class PathSmoother:
         :param equal_aspect_ratio: A bool defining whether the axes should have equal aspect ratio
         """
         assert self.atom_positions is not None and self.n_interaction_frames is not None
-        plot_atom_trajectories(self.atom_positions, self.n_interaction_frames, equal_aspect_ratio)
+        plot_atom_trajectories(
+            self.atom_positions, self.n_interaction_frames, equal_aspect_ratio
+        )
 
 
 def get_uf_atoms_and_frames(user_forces: np.ndarray) -> np.ndarray:
@@ -388,7 +400,12 @@ def calculate_com(atom_positions: np.ndarray, atom_masses: np.ndarray) -> np.nda
     ) / np.sum(atom_masses)
 
 
-def plot_com_trajectory(atom_positions: np.ndarray, n_frames: int, equal_aspect_ratio: bool = False) -> None:
+def plot_com_trajectory(
+    atom_positions: np.ndarray,
+    n_frames: int,
+    equal_aspect_ratio: bool = False,
+    cmap: str = "viridis",
+) -> None:
     """
[…]
