### 2025-05-15 f1663e7
Update colour map docstring
