### 2025-05-15 626221c
Add colour map option to PathSmoother class functions
