### 2025-05-15 259ed9b
Call make_plots_interactive only in interactive plots themselves, return the smoothing_plot function directly, and save the smoothing plot parameters as attributes of the class
