### 2025-05-15 fff2e8f
Remove global variables
