### 2025-05-15 1ba3964
Fix start and end value sliders of interactive plot to take values less than half of the length of the positions array being plotted
