### 2025-05-15 5690126
Change smoothing value slider to a float slider rather than an integer slider
