### 2025-05-15 7f4706f
Implement equal aspect ratio and colour map options for interactive smoothing plot
