### 2025-05-15 ffac75d
Add colour bar legend to interactive plot
