### 2025-05-15 d443dac
Add general interpolate_path function
