+++ python-libraries/nanover-server/src/nanover/smd/pathsmoothing.py
@@ -125,6 +125,7 @@ class PathSmoother:
         self.smoothing_start_index: Optional[int] = None
         self.smoothing_end_index: Optional[int] = None
 
+        self.smoothed_path: Optional[np.ndarray] = None
 
     def create_mda_universe(self):
         """
@@ -173,7 +174,9 @@ class PathSmoother:
         self.create_mda_universe()
         self.read_mda_universe_data()
 
-    def plot_com_trajectory(self, equal_aspect_ratio: bool = False, cmap: str = "viridis"):
+    def plot_com_trajectory(
+        self, equal_aspect_ratio: bool = False, cmap: str = "viridis"
+    ):
         """
         Plot the trajectory of the COM of the atoms defining the path.
 
@@ -186,7 +189,9 @@ class PathSmoother:
             self.com_positions, self.n_interaction_frames, equal_aspect_ratio, cmap
         )
 
-    def plot_atoms_trajectories(self, equal_aspect_ratio: bool = False, cmap: str = "viridis"):
+    def plot_atoms_trajectories(
+        self, equal_aspect_ratio: bool = False, cmap: str = "viridis"
+    ):
         """
         Plot the trajectories of the individual atoms defining the path.
 
@@ -199,7 +204,9 @@ class PathSmoother:
             self.atom_positions, self.n_interaction_frames, equal_aspect_ratio, cmap
         )
 
-    def create_interactive_smoothing_plot(self, equal_aspect_ratio: bool = False, cmap: str = "viridis"):
+    def create_interactive_smoothing_plot(
+        self, equal_aspect_ratio: bool = False, cmap: str = "viridis"
+    ):
         """
         Create an interactive plot to smooth the trajectory of the centre of mass of the atoms defining
         the path.
@@ -209,14 +216,21 @@ class PathSmoother:
         """
         self._make_plots_interactive()
 
-        def interactive_smoothing_plot(x_pos, y_pos, z_pos, smoothing_value, n_points, start_point, end_point):
+        def interactive_smoothing_plot(
+            x_pos, y_pos, z_pos, smoothing_value, n_points, start_point, end_point
+        ):
 
             pos_array_size = x_pos.size
             # start_point = min(int(x_pos.size)/2., start_point)
             # end_point = min(int(x_pos.size)/2., end_point)
             tck, u = splprep(
-                [x_pos[start_point:pos_array_size - end_point], y_pos[start_point:pos_array_size - end_point],
-                 z_pos[start_point:pos_array_size - end_point]], s=smoothing_value);
+                [
+                    x_pos[start_point : pos_array_size - end_point],
+                    y_pos[start_point : pos_array_size - end_point],
+                    z_pos[start_point : pos_array_size - end_point],
+                ],
+                s=smoothing_value,
+            )
             # tck, u = splprep([x_pos, y_pos, z_pos], s=smoothing_value);
             x_knots, y_knots, z_knots = splev(tck[0], tck)
             u_fine = np.linspace(0, 1, n_points)
@@ -230,20 +244,24 @@ class PathSmoother:
                 plt.close(self.fig)
 
                 self.fig = plt.figure(figsize=(8, 8))
-                self.ax = plt.axes(111, projection='3d')
-                self.scatter_curve = self.ax.scatter3D(x_fine,
-                                             y_fine,
-                                             z_fine,
-                                             c=np.linspace(0, 1, len(x_fine)),
-                                             cmap=cmap,
-                                             s=1.0)
-                self.scatter_points = self.ax.scatter3D(x_pos,
-                                              y_pos,
-                                              z_pos,
-                                              c=np.linspace(0, 1, pos_array_size),
[…]
