### 2025-05-15 d23e39d
Start cleaning up interactive smoothing plot function
