### 2025-05-15 ba18753
Save the smoothed COM trajectory and define a function to plot it
