### 2025-05-15 1cd0238
Edit docstring for plot_smoothed_com_trajectory
