### 2025-05-15 1cc8ae0
Transpose coordinates output by interpolate_path and add docstrings for interpolate_path
