### 2025-05-15 #8118 @jkrajniak
> [@jkrajniak](https://github.com/jkrajniak) Typically once you have completed your checklist, you can make your recommendation in comment. I see there are a few checklist items still open.
>
> If you think something need to improved in software or there are blocking comments, please create a new issue in the [target repository](https://github.com/IRL2/nanover-server-py) and link to those issues in comments.

I have a bit of doubts regarding this point:
> Contribution and authorship: Has the submitting author (@hjstroud) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?;

When I look into stats: https://github.com/IRL2/nanover-server-py/graphs/contributors, hjstroud is in 6th place, which comes from contributions to the repository. Does it fulfil the requirements of this point? Also, the list of authors in the paper does not entirely correspond to the list of contributors to the repository.
