### 2025-05-15 #8118 @abhishektiwari
@hjstroud can you provide a response to @jkrajniak comment on contribution and authorship.
