### 2025-05-16 8f82216
Change close_interactive_plots to make sure it resets interactive smoothing plot attributes
