### 2025-05-16 aa009df
Add function to calculate a constant speed trajectory defining the path to be taken by an SMD force
