### 2025-05-16 0c1457e
Add functionality to plot  the constant speed trajectory calculated using `plot_constant_speed_trajectory` and to save the data associated with this calculated trajectory
