### 2025-05-19 #8118 @hjstroud
Regarding the list of authors on the paper:

- From the activity graphs on the Git repository (https://github.com/IRL2/nanover-server-py/graphs/contributors), the development history is divided into two sections: pre-2023 (i.e. Narupa) and 2023 onwards (i.e. NanoVer)
- Narupa was developed by a different set of authors, and this development is documented by publications such as (https://doi.org/10.1063/1.5092590, https://doi.org/10.1126/sciadv.aat2731 and https://doi.org/10.1145/3388536.3407891), which we have cited in this paper
- NanoVer is the successor to Narupa, and the NanoVer repos (including the NanoVer Server and NanoVer iMD-VR) were made public with an MIT license in 2024 with the authorisation of the original institution
- NanoVer development has included implementing modern headset support (involving substantial updating of the codebase) and improving the quantitative rigour of the iMD protocol and its implementation, the latter of which is in part the subject of this JOSS submission
- All code development on NanoVer (from 2023 onwards) was performed by authors included in this submission to JOSS

Regarding my (@hjstroud) contributions:

- When examining contributions from 2023 onwards, I appear 3rd by commits on the contributors page of the NanoVer Server repo (https://github.com/IRL2/nanover-server-py/graphs/contributors) and 3rd by commits on the contributors page of the NanoVer Docs repo (https://github.com/IRL2/nanover-docs/graphs/contributors)
- My contributions constitute a large part of the effort to provide quantitative rigour in the simulation and data reporting, comprehensively document the software, and improve & expand the suite of tutorials, all of which advance the project further along the trajectory from a proof of concept to a scientific research tool
- In addition to code contributions, I drafted the present submission and coordinated the subsequent revisions of the other authors prior to submission
- The contributions of Mark (@Ragzouken) and Jonathan (@jbarnoud), who appear first and second by commits respectively from 2023 onwards, focus more heavily on maintenance and modernisation of the codebase to support the development of this expanded utility
