### 2025-05-21 ffb7f34
Add method to reset SMD simulation
