### 2025-05-21 3b1181c
Make "calculate_forces" private
