### 2025-05-21 be52d83
Ensure SMD force is present when attempting to perform SMD simulation
