### 2025-05-22 3fdc423
Add option to save smd force to file if desired when generating the starting structures for SMD
