### 2025-05-23 73df021
First attempt at string optimiser (very sketchy)
