+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -11,6 +11,8 @@ from openmm.app import Simulation
 
 from nanover.openmm import serializer
 
+from nanover.smd.pathsmoothing import interpolate_path
+
 
 class OpenMMSMDSimulation:
     """
@@ -691,6 +693,243 @@ class OpenMMSMDSimulationCOM(OpenMMSMDSimulation):
         self.calculate_com_trajectory()
         self._calculate_forces(self.com_positions)
         self._calculate_work_done()
+
+
+class OpenMMStringOptimiser:
+
+    """
+    Class for optimising the path defining the trajectory of an SMD force
+    during an SMD simulation using the finite temperature string method.
+    """
+
+    def __init__(self, smd_simulation: Union[OpenMMSMDSimulationAtom, OpenMMSMDSimulationCOM], **kwargs):
+
+        self.smd_simulation = smd_simulation
+        self.checkpoints: Optional[Any] = None
+        self.node_positions: Optional[np.ndarray] = None
+        self.tangent_vectors: Optional[np.ndarray] = None
+
+    def generate_nodes(self, n_nodes):
+        """
+        Generate the checkpoint files defining the starting node positions and
+        contexts for the string optimisation.
+
+        :param n_nodes: Number of nodes to generate along the SMD force path (excluding
+          the zeroth node marking the initial geometry).
+        """
+
+        # Calculate number of steps to run between nodes, discarding remainder
+        # to produce n_nodes evenly spaced nodes (except ends)
+        n_smd_steps = self.smd_simulation.smd_path.shape[0]
+        node_interval = n_smd_steps//n_nodes
+
+        checkpoints = []
+        node_positions = np.zeros((n_nodes+1, 3))
+
+        # Save initial checkpoint and force position
+        assert np.all(self.smd_simulation.current_smd_force_position == self.smd_simulation.smd_path[0])
+        checkpoints.append(self.smd_simulation.simulation.context.createCheckpoint())
+        node_positions[0, :] = self.smd_simulation.current_smd_force_position
+
+        print("Generated checkpoint 0 (initial force position)")
+
+        for i in range(n_nodes):
+
+            step_base = i * node_interval
+
+            for step in range(node_interval):
+
+                force_position_index = step + step_base
+                #print(f"Step {force_position_index}")
+
+                # Update force position index
+                self.smd_simulation.current_smd_force_position_index = force_position_index
+
+                #print(self.smd_simulation.smd_path[self.smd_simulation.current_smd_force_position_index])
+
+                # Update SMD force position
+                self.smd_simulation.update_smd_force_position()
+
+                #print(self.smd_simulation.current_smd_force_position)
+
+                # Perform single simulation step
+                self.smd_simulation.simulation.step(1)
+
+            # Save checkpoint and force position
+            checkpoints.append(self.smd_simulation.simulation.context.createCheckpoint())
+            node_positions[i+1, :] = self.smd_simulation.current_smd_force_position
+
+            # print(f"Steps completed: {force_position_index}")
+            print(f"Generated checkpoint {i+1}")
+
+        print("All checkpoints generated")
+
+        self.checkpoints = checkpoints
+        self.node_positions = node_positions
+        self.smd_simulation.reset()
+
+    def calculate_tangent_vectors(self):
+
+        assert self.node_positions is not None
+        n_nodes = self.node_positions.shape[0]
+
+        tangent_vectors = np.zeros((n_nodes, 3))
+
+        # Determine tangent vectors for all but start and end nodes
+        differences = np.diff(self.node_positions, axis=0)
+        for i in range(1, n_nodes - 1):
+            tangent_vec = differences[i-1] + differences[i]
+            normalised_tangent_vec = tangent_vec / np.linalg.norm(tangent_vec)
[…]
