### 2025-05-23 ebdb43b
Sketch out ideas for appropriate tests
