### 2025-05-26 c8d4d25
Add attribute to record history of node updates
