### 2025-05-27 fcf5240
refactor: extract read_buffer/read_header
