### 2025-05-27 66904bb
refactor: iter_buffers with io position
