### 2025-05-27 9a38d09
refactor: file reader
