### 2025-05-27 61cb3eb
perf improvements + alternate universes function
