### 2025-05-27 c98b5ba
Add basic MEP optimiser (not sure it actually works yet)
