### 2025-05-28 851e534
fix implementation for irregular recordings
