+++ python-libraries/nanover-server/tests/test_core/recording/test_reading.py
@@ -11,9 +11,6 @@ from nanover.recording.reading import (
     split_by_simulation_counter,
 )
 
-from nanover.protocol.trajectory import GetFrameResponse
-from nanover.protocol.state import StateUpdate
-
 EXAMPLES_PATH = Path(__file__).parent
 RECORDING_PATH_TRAJ = EXAMPLES_PATH / "nanotube-example-recording.traj"
 RECORDING_PATH_STATE = EXAMPLES_PATH / "nanotube-example-recording.state"
