+++ python-libraries/nanover-server/src/nanover/mdanalysis/universe.py
@@ -21,9 +21,9 @@ Facilities to read a NanoVer trajectory recording into an MDAnalysis Universe.
 """
 
 import warnings
-from itertools import takewhile, chain, islice
+from itertools import islice
 from os import PathLike
-from typing import NamedTuple, Type, Callable, Iterable
+from typing import NamedTuple, Type, Callable
 
 from MDAnalysis import Universe
 from MDAnalysis.coordinates.base import ProtoReader
@@ -65,10 +65,7 @@ from nanover.trajectory.frame_data import (
 
 from nanover.recording.reading import (
     iter_trajectory_recording,
-    iter_trajectory_with_elapsed_integrated,
     advance_to_first_particle_frame,
-    advance_to_first_coordinate_frame,
-    split_by_simulation_counter,
     MessageRecordingReader,
     read_header,
     iter_buffers,
@@ -98,7 +95,11 @@ KEY_TO_ATTRIBUTE = {
 }
 
 
-def universes_from_recording2(*, traj: PathLike[str]):
+def universes_from_recording(*, traj: PathLike[str]):
+    """
+    Decompose a NanoVer trajectory recording into an mdanalysis Universe for each session of simulation (determined
+    by frame_index resets).
+    """
     frame_offsets: list[int] = []
     universes: list[Universe] = []
     first_particle_frame = FrameData()
@@ -108,9 +109,8 @@ def universes_from_recording2(*, traj: PathLike[str]):
         if not frame_offsets:
             return
 
-        reader = NanoverReaderFile(traj)
+        reader = NanoverReader(traj)
         reader.reader.message_offsets = list(frame_offsets)
-        reader.replace_reader(reader.reader)
         frame_offsets.clear()
 
         try:
@@ -133,35 +133,6 @@ def universes_from_recording2(*, traj: PathLike[str]):
             frame_offsets.append(entry.offset)
 
     make_universe()
-
-    return universes
-
-
-def universes_from_recording(*, traj: PathLike[str]):
-    """
-    Decompose a NanoVer trajectory recording into an mdanalysis Universe for each session of simulation (determined
-    by simulation counter).
-    """
-    universes = []
-
-    for session in split_by_simulation_counter(traj=traj):
-        # universe from first frame of session containing initial positions
-        first_positions_frame = next(
-            frame
-            for (elapsed, frame, state) in session
-            if PARTICLE_POSITIONS in frame
-            and PARTICLE_COUNT in frame
-            and frame.particle_count > 0
-        )
-        universe = frame_data_to_mdanalysis(first_positions_frame)
-        # integrate time elapsed into frames
-        for elapsed, frame, state in session:
-            frame.values["elapsed"] = elapsed
-        # trajectory from all frames of session
-        universe.trajectory = NanoverFramesReader(
-            frame for (elapsed, frame, state) in session
-        )
-        universes.append(universe)
 
     return universes
 
@@ -231,31 +202,53 @@ class NanoverParser(TopologyReaderBase):
             )
 
 
-class NanoverFramesReader(ProtoReader):
+def reconfigure_frame_reader(reader: MessageRecordingReader):
+    first_frame = FrameData()
+    for i, entry in enumerate(reader):
+        message = buffer_to_frame_message(entry.buffer)
+        first_frame.raw.MergeFrom(message.frame)
+        if PARTICLE_POSITIONS in first_frame:
+            reader.message_offsets = reader.message_offsets[i:]
+            return first_frame
+
+
+def trim_frame_reader(reader: MessageRecordingReader):
+    for i, entry in enumerate(reader):
+        message = buffer_to_frame_message(entry.buffer)
+        if message.frame_index == 0 and i > 0:
[…]
