### 2025-05-30 99d654c
docstrings
