### 2025-06-02 55ed7a4
support constructing traj reader with existing indexed file
