### 2025-06-02 d9d4711
simplify tests
