### 2025-06-06 d360108
singular universe method + don't fail to convert bonds if there's no order information
