### 2025-06-06 #8118 @samuelmurail
Hi @abhishektiwari,

I am fine with the paper and deposit; it was just hard to find all the 'Community Guidelines', which are present but scattered across different locations. Similarly, I could not pass the test when trying to launch an interactive simulation with Jupyter because some libraries were missing (I probably did not execute all the developers' installation steps).

I would suggest citing the MDDriver code (https://github.com/LBT-CNRS/MDDriver), which also proposes interactive MD simulations with VR (though probably not with multiple users).

I should also mention that I don't have a VR headset, so I could not test the VR part — only the Jupyter notebook and tutorial.

Apart from these minor suggestions, I approve the manuscript and deposit.
