### 2025-06-09 3655e3e
fix bug with single frame trajectories
