+++ python-libraries/nanover-server/src/nanover/mdanalysis/converter.py
@@ -327,6 +327,8 @@ def _add_bonds_to_mda(u: Universe, frame: FrameData):
     :param frame: NanoVer :class:`FrameData`.
     """
     try:
+        # TODO: why does mypy hate this?
+        # type: ignore
         bonds = [(bond[0], bond[1]) for bond in frame.bond_pairs]
     except MissingDataError:
         return
