### 2025-06-09 8c6f368
Improve recordings performance and utilities (#410)
refactor: mdanalysis reader for nanover recordings now reads on demand from file instead of copying everything to memory. code for reading recordings also improved.
