+++ python-libraries/nanover-server/src/nanover/recording/utilities.py
@@ -1,0 +1,188 @@
+import math
+from dataclasses import dataclass
+from os import PathLike
+from typing import Optional, BinaryIO, Callable
+
+from nanover.protocol.trajectory import GetFrameResponse
+from nanover.protocol.state import StateUpdate
+from nanover.recording.reading import (
+    MessageRecordingReader,
+    buffer_to_frame_message,
+    buffer_to_state_message,
+)
+from nanover.recording.writing import write_entry, write_header
+from nanover.state.state_dictionary import StateDictionary
+from nanover.state.state_service import (
+    dictionary_change_to_state_update,
+    state_update_to_dictionary_change,
+)
+from nanover.trajectory import FrameData
+from nanover.trajectory.frame_data import SIMULATION_COUNTER
+from nanover.utilities.change_buffers import DictionaryChange
+
+
+@dataclass(kw_only=True)
+class FrameRecordingEvent:
+    timestamp: int
+    message: GetFrameResponse
+    current_frame: FrameData
+
+
+@dataclass(kw_only=True)
+class StateRecordingEvent:
+    timestamp: int
+    message: StateUpdate
+    current_state: dict
+
+
+@dataclass(kw_only=True)
+class RecordingEvent:
+    timestamp: int
+    frame_event: Optional[FrameRecordingEvent]
+    state_event: Optional[StateRecordingEvent]
+
+
+def split_on_frame_reset(event: RecordingEvent):
+    return event.frame_event is not None and event.frame_event.message.frame_index == 0
+
+
+def split_on_sim_counter_change(event: RecordingEvent):
+    return (
+        event.frame_event is not None
+        and SIMULATION_COUNTER in event.frame_event.message.frame.values
+    )
+
+
+def split_recording(
+    *,
+    traj: Optional[PathLike[str]] = None,
+    state: Optional[PathLike[str]] = None,
+    split_predicate: Callable[[RecordingEvent], bool] = split_on_frame_reset,
+):
+    """
+    :param traj: Path of a NanoVer trajectory recording
+    :param state: Path of a NanoVer state recording
+    :param split_predicate: Predicate function that takes information about the next frame and returns True if the
+    recording should split at this point.
+    """
+    split_count = 0
+    current_base_timestamp = 0
+    current_traj_out: Optional[BinaryIO] = None
+    current_state_out: Optional[BinaryIO] = None
+
+    def close_all():
+        if current_state_out is not None:
+            current_traj_out.close()
+        if current_state_out is not None:
+            current_state_out.close()
+
+    def open_all():
+        nonlocal current_traj_out, current_state_out
+        if traj is not None:
+            current_traj_out = open(f"{traj}--SPLIT--{split_count}.traj", "wb")
+            write_header(current_traj_out)
+        if state is not None:
+            current_state_out = open(f"{state}--SPLIT--{split_count}.state", "wb")
+            write_header(current_state_out)
+
+    def dict_to_state_update(dict):
+        change = DictionaryChange(updates=dict)
+        state = dictionary_change_to_state_update(change)
+        return state
+
+    try:
+        open_all()
+
+        prev_frame = None
+        prev_state = None
+
+        for event in iter_recording_max(traj=traj, state=state):
+            if event.frame_event is not None:
+                prev_frame = event.frame_event
+            if event.frame_event is not None:
+                prev_state = event.state_event
+
+            if split_predicate(event):
+                close_all()
+                split_count += 1
+                current_base_timestamp = event.timestamp
+                open_all()
+                if prev_frame is not None:
+                    frame = GetFrameResponse(
+                        frame_index=0, frame=prev_frame.current_frame.raw
+                    )
+                    write_entry(current_traj_out, 0, frame)
+                if prev_state is not None:
+                    state_update = dict_to_state_update(prev_state.current_state)
[…]
