+++ joss-paper-2025/paper.bib
@@ -882,6 +882,21 @@
 	pages = {108932},
 }
 
+% MDDriver paper (used with UnityMol for iMD-VR)
+@article{delalande_complex_2009,
+	title = {Complex molecular assemblies at hand via interactive simulations},
+	volume = {30},
+	issn = {1096-987X},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21235},
+	doi = {10.1002/jcc.21235},
+	abstract = {Studying complex molecular assemblies interactively is becoming an increasingly appealing approach to molecular modeling. Here we focus on interactive molecular dynamics (IMD) as a textbook example for interactive simulation methods. Such simulations can be useful in exploring and generating hypotheses about the structural and mechanical aspects of biomolecular interactions. For the first time, we carry out low-resolution coarse-grain IMD simulations. Such simplified modeling methods currently appear to be more suitable for interactive experiments and represent a well-balanced compromise between an important gain in computational speed versus a moderate loss in modeling accuracy compared to higher resolution all-atom simulations. This is particularly useful for initial exploration and hypothesis development for rare molecular interaction events. We evaluate which applications are currently feasible using molecular assemblies from 1900 to over 300,000 particles. Three biochemical systems are discussed: the guanylate kinase (GK) enzyme, the outer membrane protease T and the soluble N-ethylmaleimide-sensitive factor attachment protein receptors complex involved in membrane fusion. We induce large conformational changes, carry out interactive docking experiments, probe lipid–protein interactions and are able to sense the mechanical properties of a molecular model. Furthermore, such interactive simulations facilitate exploration of modeling parameters for method improvement. For the purpose of these simulations, we have developed a freely available software library called MDDriver. It uses the IMD protocol from NAMD and facilitates the implementation and application of interactive simulations. With MDDriver it becomes very easy to render any particle-based molecular simulation engine interactive. Here we use its implementation in the Gromacs software as an example. © 2009 Wiley Periodicals, Inc. J Comput Chem, 2009},
+	number = {15},
+	journal = {Journal of Computational Chemistry},
+	author = {Delalande, Olivier and Férey, Nicolas and Grasseau, Gilles and Baaden, Marc},
+	year = {2009},
+	pages = {2375--2387},
+}
+
 % Subtle sensing CHI paper
 @inproceedings{roebuck_williams_subtle_2020,
 	address = {New York, NY, USA},
