+++ joss-paper-2025/paper.md
@@ -71,7 +71,7 @@ The family of simulation methods encompassed by the term molecular dynamics (MD)
 
 Interactive molecular dynamics (iMD) is a nonequilibrium enhanced sampling method that allows researchers to adaptively bias molecular simulations on-the-fly in real time [@stone_system_2001; @rapaport_interactive_1997; @oconnor_sampling_2018; @oconnor_interactive_2019]. In iMD, researchers can tailor biasing forces in real time in response to the evolving system dynamics, eliminating the need to pre-program perturbations. A number of programs have implementations of iMD, including for example TeraChem [@luehr_ab_2015], SCINE [@weymuth_immersive_2021], and NAMD/VMD [@stone_system_2001], which enable researchers to interact via haptic interfaces with molecular systems visualised on 2-D screens. Fewer programs are available that enable researchers to interact with 3-D MD simulations in a natively 3-D environment, which is especially important for exploring complex molecular structural transformations [@oconnor_sampling_2018]. One example is ExaViz, where users can interact with molecules in a CAVE-like virtual environment [@dreher_exaviz_2014].
 
-Virtual reality (VR) provides a natural interface for visualising molecular systems by mapping the 3-D simulation space of the molecular system to a 3-D virtual environment that the researcher can inhabit. While several programs already enable molecular visualisation in VR [@doutreligne_unitymol_2014; @bennie_virtual_2023; @pettersen_ucsf_2021; @ozvoldik_yasara_2023; @cassidy_proteinvr_2020; @cortes_rodriguez_molecularwebxr_2025], few support the combination of VR-enabled 3-D visualisation with real-time iMD. Performing iMD in VR (iMD-VR) allows users to interact directly with molecular simulations in their native 3-D environment, enabling researchers to reach out and 'touch' molecules as if they were tangible objects [@roebuck_williams_subtle_2020; @roebuck_williams_measuring_2024]. Many studies have demonstrated the utility of iMD-VR for research applications, in areas spanning protein-ligand binding, [@deeks_interactive_2020; @deeks_interactive_docking_2020; @henry_chan_discovery_2021; @walters_emerging_2022], protein conformational dynamics [@juarez_jimenez_combining_2020], machine-learning potential energy surfaces [@amabilino_training_2019; @amabilino_training_2020], discovering reaction networks [@shannon_exploring_2021], materials science and catalysis [@crossley_lewis_interactive_2023], and chemistry & biochemistry education [@bennie_teaching_2019].
[…]
