### 2025-06-10 1665557
Start writing tests for OpenMM SMD classes (test loading from simulation and xml file path)
