+++ python-libraries/nanover-server/tests/test_smd/test_openmm.py
@@ -38,23 +38,164 @@ Things to test:
 - smd_com_force works as expected
 - smd_single_atom_force works as expected
 """
+import pytest
+from contextlib import contextmanager
 
+import numpy as np
 import openmm as mm
-
-from nanover.smd.openmm import *
-
-from ..test_openmm.simulation_utils import (
-    basic_system,
-    basic_simulation,
-    basic_simulation_with_imd_force,
-    BASIC_SIMULATION_POSITIONS,
-    empty_imd_force,
-    assert_basic_simulation_topology,
-    single_atom_system,
-    single_atom_simulation,
-    single_atom_simulation_with_imd_force,
-    ARGON_SIMULATION_POSITION,
-    assert_single_atom_simulation_topology,
+from openmm import app
+from openmm.unit import (
+    kelvin,
+    picosecond,
+    femtosecond,
+    nanometer,
 )
 
+from nanover.openmm import serializer
+from nanover.smd.openmm import *
+
 # Very basic thing to test entire class as it would be used: tutorial notebook that can be tested
+
+BASIC_SIMULATION_BOX_VECTORS = [[50, 0, 0], [0, 50, 0], [0, 0, 50]]
+BASIC_SIMULATION_POSITIONS = [
+    # First residue
+    [0, 0, 0],  # C
+    [5.288, 1.610, 9.359],  # H
+    [2.051, 8.240, -6.786],  # H
+    [-10.685, -0.537, 1.921],  # H
+    # Second residue, copied from the first but shifted
+    # by 5 nm along the Z axis
+    [0, 0, 5],  # C
+    [5.288, 1.610, 14.359],  # H
+    [2.051, 8.240, -1.786],  # H
+    [-10.685, -0.537, 6.921],  # H
+]
+ARGON_SIMULATION_POSITION = [[0.0, 0.0, 0.0]]
+
+TEST_SMD_SINGLE_INDEX = np.array(0)
+TEST_SMD_MULTIPLE_INDICES = np.array([0, 1, 2, 3])
+TEST_SMD_PATH = np.array([np.linspace(0.0,1.0,101), np.zeros(101), np.zeros(101)]).transpose()
+TEST_SMD_FORCE_CONSTANT = 3011.0
+
+TEST_SMD_INDEX_OPTIONS = [TEST_SMD_SINGLE_INDEX, TEST_SMD_MULTIPLE_INDICES]
+EXPECTED_SMD_SIMULATION_TYPES = [OpenMMSMDSimulationAtom, OpenMMSMDSimulationCOM]
+
+def build_basic_system():
+    periodic_box_vector = BASIC_SIMULATION_BOX_VECTORS
+    system = mm.System()
+    system.setDefaultPeriodicBoxVectors(*periodic_box_vector)
+    system.addParticle(mass=12)
+    system.addParticle(mass=1)
+    system.addParticle(mass=1)
+    system.addParticle(mass=1)
+    system.addParticle(mass=12)
+    system.addParticle(mass=1)
+    system.addParticle(mass=1)
+    system.addParticle(mass=1)
+    return system
+
+
+def build_basic_topology() -> app.Topology:
+    topology = app.Topology()
+    carbon = app.Element.getBySymbol("C")
+    hydrogen = app.Element.getBySymbol("H")
+    chain = topology.addChain()
+    residue = topology.addResidue(name="METH1", chain=chain)
+    atom_c1 = topology.addAtom(element=carbon, name="C1", residue=residue)
+    atom_h2 = topology.addAtom(element=hydrogen, name="H2", residue=residue)
+    atom_h3 = topology.addAtom(element=hydrogen, name="H3", residue=residue)
+    atom_h4 = topology.addAtom(element=hydrogen, name="H4", residue=residue)
+    topology.addBond(atom_c1, atom_h2)
[…]
