### 2025-06-10 2f9f7c6
Formatting with Black
