+++ python-libraries/nanover-server/src/nanover/smd/pathsmoothing.py
@@ -91,8 +91,6 @@ class PathSmoother:
                 "interfaces."
             )
 
-
-
     def __init__(self):
         self.filename: Optional[Union[str, PathLike]] = None
         self.com_positions: Optional[np.ndarray] = None
@@ -231,7 +229,9 @@ class PathSmoother:
             pos_array_size = x_pos.size
             original_u_values = np.linspace(0, 1, pos_array_size)
 
-            interpolated_path, interpolated_u_values = interpolate_path(x_pos, y_pos, z_pos, smoothing_value, n_points, start_point, end_point)
+            interpolated_path, interpolated_u_values = interpolate_path(
+                x_pos, y_pos, z_pos, smoothing_value, n_points, start_point, end_point
+            )
 
             # Save current view if plot already exists
             if self.fig is None or self.ax is None:
@@ -242,7 +242,9 @@ class PathSmoother:
                 self.fig = plt.figure(figsize=(8, 8))
                 self.ax = plt.axes(111, projection="3d")
                 self.scatter_curve = self.ax.scatter3D(
-                    interpolated_path[:, 0], interpolated_path[:, 1], interpolated_path[:, 2],
+                    interpolated_path[:, 0],
+                    interpolated_path[:, 1],
+                    interpolated_path[:, 2],
                     c=interpolated_u_values,
                     cmap=cmap,
                     s=1.0,
@@ -281,13 +283,17 @@ class PathSmoother:
                 self.scatter_curve.remove()
                 self.scatter_points.remove()
                 self.scatter_curve = self.ax.scatter3D(
-                    interpolated_path[:, 0], interpolated_path[:, 1], interpolated_path[:, 2],
+                    interpolated_path[:, 0],
+                    interpolated_path[:, 1],
+                    interpolated_path[:, 2],
                     c=interpolated_u_values,
                     cmap=cmap,
                     s=1.0,
                 )
                 self.scatter_points = self.ax.scatter3D(
-                    x_pos, y_pos, z_pos,
+                    x_pos,
+                    y_pos,
+                    z_pos,
                     c=original_u_values,
                     cmap=cmap,
                     s=50.0,
@@ -366,11 +372,17 @@ class PathSmoother:
             self.constant_speed_com_trajectory, n_points, equal_aspect_ratio, cmap
         )
 
-    def calculate_constant_speed_trajectory(self, desired_speed_nm_ps: float, timestep_ps: float, rel_tolerance: float = 1e-5, max_iterations: int = 10):
+    def calculate_constant_speed_trajectory(
+        self,
+        desired_speed_nm_ps: float,
+        timestep_ps: float,
+        rel_tolerance: float = 1e-5,
+        max_iterations: int = 10,
+    ):
         """
         Calculate the coordinates that define a constant speed trajectory along a smoothed
         COM trajectory, given the desired speed in nm ps-1 and timestep of the simulation in ps.
-        
+
         :param desired_speed_nm_ps: The desired speed in nm ps-1
         :param timestep_ps: The timestep in ps
         :param rel_tolerance: The relative tolerance in the errors of the speed and distance between points
@@ -379,30 +391,56 @@ class PathSmoother:
 
         assert self.smoothed_com_trajectory is not None
         path_length = calculate_trajectory_length(self.smoothed_com_trajectory)
-        n_points_required = int(np.ceil(path_length / (desired_speed_nm_ps * timestep_ps)))# + 1
-        refined_path, _ = interpolate_path(self.smoothed_com_trajectory[:,0],self.smoothed_com_trajectory[:,1],self.smoothed_com_trajectory[:,2], 0.0, n_points=n_points_required)
+        n_points_required = int(
+            np.ceil(path_length / (desired_speed_nm_ps * timestep_ps))
+        )  # + 1
+        refined_path, _ = interpolate_path(
[…]
