### 2025-06-11 #8118 @abhishektiwari
@hjstroud can you check following,

1. Is it "Technoloxías Intelixentes" or "Tecnoloxías Intelixentes"
2. Is it 𝒪(10^9) atoms or O(10^9) or "on the order of 10^9?
