### 2025-06-11 #8118 @samuelmurail
> Of course [@abhishektiwari](https://github.com/abhishektiwari) , thank you for the feedback [@samuelmurail](https://github.com/samuelmurail) !
>
> To our best knowledge, MDDriver is a middleware that facilitates interactive molecular dynamics by acting as an interface between a molecular visualisation program and a simulation engine (e.g. VMD/MDDriver/GROMACS), but does not facilitate iMD-VR itself, and as far as we know is not capable of multi-user iMD. A [recent demonstration](https://conferences.mpi-inf.mpg.de/dagm/2020/wp-content/uploads/2020/09/abstract1005.pdf) combined MDDriver with UnityMol to facilitate iMD-VR, but to our best knowledge there have been no research applications of this protocol, and hence we have not cited it as research software with this capability. I propose a slight amendment to the sentence
>
> "While several programs already enable molecular visualisation in VR, few support the combination of VR-enabled 3-D visualisation with real-time iMD."
>
> to
>
> "While several programs already enable molecular visualisation in VR, few support the combination of VR-enabled 3-D visualisation with real-time iMD, with notable examples including Narupa and UnityMol."
>
> and then citing both UnityMol and MDDriver when referencing UnityMol in this sentence. Would this be sufficient?

I'm fine with this. Thanks
