### 2025-06-11 bdd94a1
Add test to check that OpenMMSMDSimulation returns the correct subclass based on the indices passed to it
