### 2025-06-12 b07dfa3
Add test to check that PBCs of the simulation are respected by the added force
