### 2025-06-12 0ab44cb
Add tests to check that the SMD forces attach to the correct atom/atoms
