### 2025-06-12 4cfcf71
Minor edits to delete properties on reset and use np.array_equal(...) in appropriate functions
