### 2025-06-12 8bd7d88
Add a function to check that resetting the class is equivalent to creating a new version of the class, with all variables reset to their original state correctly
