### 2025-06-12 d1feee3
Add function to check that the SMD force is added to the system correctly
