### 2025-06-12 072c238
Add a function that tests that the SMD force can be removed correctly from the system
