### 2025-06-12 0321857
Add test to check that update_smd_force_position() correctly updates the position of the SMD force, both in the class and in the OpenMM system
