### 2025-06-12 3c4a779
Add test to check that equilibration performed around non-initial SMD force position throws an error
