### 2025-06-12 a73e3b3
Formatting with Black
