### 2025-06-13 36bb0f3
Add assert to check that force constant global parameter is still present (should still be there at the moment) and TODO about removing it
