### 2025-06-18 #8118 @hjstroud
@abhishektiwari since the present submission, we have updated some aspects of the codebase (mainly fixes to issues). When we make a release of the software, should we use the latest version reviewed in this submission or the latest version of the software itself according to the current state of the main branch?
