### 2025-06-24 39726e9
Add tests to check that the SMD forces can be correctly loaded from xml files containing an SMD force
