### 2025-06-25 #8118 @abhishektiwari
@hjstroud  Version number you shared is for Zenodo. We need you to make a tagged release of repository and archive, and report the version number and archive DOI. See example [here](https://github.com/bempp/kifmm/releases/tag/v2.1.0) and [here](https://zenodo.org/records/15696916). Hope this makes sense.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
