### 2025-06-25 3f737b6
Add tests to check that the SMD forces can be correctly loaded from xml files saved without an SMD force
