### 2025-06-25 b9f86c1
Check off another test that is already checked when running `check_for_existing_smd_force()`
