### 2025-06-25 ef75c33
Only write SMD force expressions once explicitly
