+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -11,6 +11,10 @@ from openmm.app import Simulation
 
 from nanover.openmm import serializer
 
+SMD_FORCE_EXPRESSION_ATOM_NONPERIODIC = "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)"
+SMD_FORCE_EXPRESSION_ATOM_PERIODIC = "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2"
+SMD_FORCE_EXPRESSION_COM = "0.5 * smd_k * pointdistance(x1, y1, z1, x0, y0, z0)^2"
+
 
 class OpenMMSMDSimulation:
     """
@@ -539,9 +543,9 @@ class OpenMMSMDSimulationAtom(OpenMMSMDSimulation):
             assert type(smd_force) == CustomExternalForce
             assert (
                 smd_force.getEnergyFunction()
-                == "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2"
+                == SMD_FORCE_EXPRESSION_ATOM_PERIODIC
                 or smd_force.getEnergyFunction()
-                == "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)"
+                == SMD_FORCE_EXPRESSION_ATOM_NONPERIODIC
             )
             assert smd_force.getNumParticles() == 1
             assert force_constant == self.smd_force_constant
@@ -610,7 +614,7 @@ class OpenMMSMDSimulationCOM(OpenMMSMDSimulation):
             assert type(smd_force) == CustomCentroidBondForce
             assert (
                 smd_force.getEnergyFunction()
-                == "0.5 * smd_k * pointdistance(x1, y1, z1, x0, y0, z0)^2"
+                == SMD_FORCE_EXPRESSION_COM
             )
             assert smd_force.getNumGroups() == 1
             assert force_constant == self.smd_force_constant
@@ -713,7 +717,7 @@ def smd_com_force(force_constant: float, uses_pbcs: bool):
 
     # TODO: Check that this works with PBCs correctly - may need to amend setUsesPBCs line...
     smd_force = CustomCentroidBondForce(
-        1, "0.5 * smd_k * pointdistance(x1, y1, z1, x0, y0, z0)^2"
+        1, SMD_FORCE_EXPRESSION_COM
     )
     smd_force.addGlobalParameter("smd_k", force_constant)
     smd_force.addPerBondParameter("x0")
@@ -738,11 +742,11 @@ def smd_single_atom_force(force_constant: float, uses_pbcs: bool):
     """
     if uses_pbcs:
         smd_force = CustomExternalForce(
-            "0.5 * smd_k * periodicdistance(x, y, z, x0, y0, z0)^2"
+            SMD_FORCE_EXPRESSION_ATOM_PERIODIC
         )
     else:
         smd_force = CustomExternalForce(
-            "0.5 * smd_k * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)"
+            SMD_FORCE_EXPRESSION_ATOM_NONPERIODIC
         )
     smd_force.addGlobalParameter("smd_k", force_constant)
     smd_force.addPerParticleParameter("x0")
