### 2025-06-25 c9ec411
Add test to check the structure files generated by `generate_starting_structures()`
