### 2025-06-25 2d6eca0
Simplify test that checks`generate_starting_structures()`
