### 2025-06-25 0e5b92e
Remove TODO about checking PBCs in `smd_com_force`
