### 2025-06-25 4b82bda
Rename _calculate_smd_forces and add test to check it works as expected
