### 2025-06-25 2b1a538
Add less uniform parameter for `test_generate_starting_structures` and add docstrings
