### 2025-06-25 152b65b
Add less uniform parameter for `test_generate_starting_structures` and add docstrings
