### 2025-06-25 #8118 @hjstroud
@abhishektiwari We have now created a release for the version of the software reviewed in this JOSS submission (v0.1.2797), which you should be able to see [here](https://github.com/IRL2/nanover-server-py/releases/tag/0.1.2797). We didn't do the latest version (0.1.2806) because that includes some changes made outside of this review (i.e. other refactoring/additions). This should mean that the version number and DOI archive above are valid, i.e.:

- Version number: 0.1.2797
- DOI on Zenodo: [10.5281/zenodo.15691916](https://doi.org/10.5281/zenodo.15691916)

Is this all ok now?
