### 2025-06-25 98113ad
Add test to check that `_calculate_work_done()` behaves as expected
