### 2025-06-25 fecfeb9
Add TODO about generalising work done test to check work done along non-linear paths
