### 2025-06-25 135b328
Test single index and multiple indices parameters when checking the smd_simulation_forces and smd_simulation_work_done
