### 2025-06-25 3f77672
Add test to check that data saved in save_smd_simulation_data saves to the correct place, can be loaded correctly and takes the same values after loading as before saving.
