### 2025-06-25 2725705
Add test to check that data saved in `save_general_smd_data` saves to the correct place, can be loaded correctly and takes the same values after loading as before saving.
