### 2025-06-26 6c48afe
Move calculation of COM to a function outside of class
