### 2025-06-26 b819da0
Add test systems and test to check that COMs are correctly calculated using `calculate_com()`
