+++ python-libraries/nanover-server/tests/test_smd/test_openmm.py
@@ -31,7 +31,8 @@ Things to test:
 - Simulation data is saved in the correct format to the correct location, and can be
   subsequently loaded back into python correctly [√]
 - General SMD data is saved in the correct format to the correct location, and can be
-  subsequently loaded back into python correctly []
+  subsequently loaded back into python correctly [√]
+- The COM of a specified group of atoms is correctly calculated [√]
 - For OpenMMSMDSimulationCOM, the COM of the specified atoms is correctly calculated []
 - For OpenMMSMDSimulationCOM, the trajectory of the COM of the specified atoms is
   correctly calculated []
@@ -75,12 +76,56 @@ BASIC_SIMULATION_POSITIONS = [
 ]
 ARGON_SIMULATION_POSITION = [[0.0, 0.0, 0.0]]
 
+# Test parameters for OpenMMSMDSimulation
 TEST_SMD_SINGLE_INDEX = np.array(0)
 TEST_SMD_MULTIPLE_INDICES = np.array([0, 1, 2, 3])
 TEST_SMD_PATH = np.array(
     [np.linspace(0.05, 1.05, 101), np.zeros(101), np.zeros(101)]
 ).transpose()
 TEST_SMD_FORCE_CONSTANT = 3011.0
+
+# Test systems for COM calculations, formatted as (positions, masses, expected COM)
+TEST_COM_TWO_ATOMS = (
+    np.array([[-1.0, 0.0, 0.0], [1.0, 0.0, 0.0]]),
+    np.array([1.0, 1.0]),
+    np.array([0.0, 0.0, 0.0]),
+)
+TEST_COM_METHANE = (
+    np.array(
+        [
+            [1.0, 1.0, 0.0],
+            [0.0, 1.0, -1.0 / np.sqrt(2)],
+            [2.0, 1.0, -1.0 / np.sqrt(2)],
+            [1.0, 0.0, 1.0 / np.sqrt(2)],
+            [1.0, 2.0, 1.0 / np.sqrt(2)],
+        ]
+    ),
+    np.array([12.01, 1.00, 1.00, 1.00, 1.00]),
+    np.array([1.0, 1.0, 0.0]),
+)
+TEST_COM_CIRCLE = (
+    np.array(
+        [[np.cos(i * (np.pi / 4)), 0.0, np.sin(i * (np.pi / 4))] for i in range(8)]
+    ),
+    np.array([2.0, 1.05, 2.0, 1.05, 2.0, 1.05, 2.0, 1.05]),
+    np.array([0.0, 0.0, 0.0]),
+)
+TEST_COM_CUBE = (
+    np.array(
+        [
+            [0.0, 1.0, 2.0],
+            [0.0, 1.0, 4.0],
+            [0.0, 3.0, 2.0],
+            [0.0, 3.0, 4.0],
+            [2.0, 1.0, 2.0],
+            [2.0, 1.0, 4.0],
+            [2.0, 3.0, 2.0],
+            [2.0, 3.0, 4.0],
+        ]
+    ),
+    np.array([6.0, 6.0, 6.0, 6.0, 6.0, 6.0, 6.0, 6.0]),
+    np.array([1.0, 2.0, 3.0]),
+)
 
 
 def build_basic_system():
@@ -769,7 +814,7 @@ def test_calculate_smd_forces(position_shifts, indices):
         TEST_SMD_FORCE_CONSTANT,
     )
     smd_sim._calculate_smd_forces(test_positions)
-    assert np.allclose(smd_sim.smd_simulation_forces, expected_forces)
+    assert np.allclose(smd_sim.smd_simulation_forces, expected_forces, atol=1e-16)
 
 
 @pytest.mark.parametrize(
@@ -802,7 +847,11 @@ def test_calculate_work_done(position_shifts, indices):
[…]
