### 2025-06-26 93bcb2d
Add TODO to change the calculate COM function to handle PBCs correctly
