### 2025-06-26 3df8aff
Create tests for the `_calculate_com` and `_calculate_com_trajectory` functions used by the OpenMMSMDSimulationCOM class
