### 2025-06-26 c5cd2aa
Minor edit to descriptions of functions
