### 2025-06-26 e3951c3
Add tests to check that the functions that construct the SMD forces return forces with the expected properties
