### 2025-06-26 b36b9e3
Refactor classmethods `from_simulation` and `from_xml_path` to remove code duplication, and format with Black
