+++ python-libraries/nanover-server/src/nanover/smd/openmm.py
@@ -12,10 +12,10 @@ from openmm.app import Simulation
 from nanover.openmm import serializer
 
 SMD_FORCE_CONSTANT_PARAMETER_NAME = "smd_k"
-SMD_FORCE_EXPRESSION_ATOM_NONPERIODIC = f"0.5 * {SMD_FORCE_CONSTANT_PARAMETER_NAME} * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)"
-SMD_FORCE_EXPRESSION_ATOM_PERIODIC = (
-    f"0.5 * {SMD_FORCE_CONSTANT_PARAMETER_NAME} * periodicdistance(x, y, z, x0, y0, z0)^2"
+SMD_FORCE_EXPRESSION_ATOM_NONPERIODIC = (
+    f"0.5 * {SMD_FORCE_CONSTANT_PARAMETER_NAME} * ((x-x0)^2 + (y-y0)^2 + (z-z0)^2)"
 )
+SMD_FORCE_EXPRESSION_ATOM_PERIODIC = f"0.5 * {SMD_FORCE_CONSTANT_PARAMETER_NAME} * periodicdistance(x, y, z, x0, y0, z0)^2"
 SMD_FORCE_EXPRESSION_COM = f"0.5 * {SMD_FORCE_CONSTANT_PARAMETER_NAME} * pointdistance(x1, y1, z1, x0, y0, z0)^2"
 
 
@@ -43,8 +43,8 @@ class OpenMMSMDSimulation:
         Construct the SMD simulation from an existing OpenMM simulation.
 
         :param simulation: An existing OpenMM Simulation
-        :param smd_atom_indices: The indices of the atoms to which the SMD force
-          should be applied
+        :param smd_atom_indices: A NumPy array of indices of the atoms to which the SMD force
+          should be applied (0-D for single atom, 1-D for COM)
         :param smd_path: A NumPy array of coordinates defining the path that the
           SMD force will take during the SMD simulation
         :param smd_force_constant: The force constant of the SMD force
@@ -60,27 +60,12 @@ class OpenMMSMDSimulation:
 
         sim.name = name
         sim.simulation = simulation
+
         # Check if simulation employs periodic boundary conditions
         sim._sim_uses_pbcs = sim.simulation.system.usesPeriodicBoundaryConditions()
 
-        sim.smd_atom_indices = smd_atom_indices
-        sim.smd_path = smd_path
-        sim.smd_force_constant = smd_force_constant
-        sim.n_smd_atom_indices = sim.smd_atom_indices.size
-        sim.define_smd_simulation_atom_positions_array()
-
-        # Create SMD force and add it to the system
-        sim.current_smd_force_position = sim.smd_path[0]
-        sim.current_smd_force_position_index = 0
-
-        # Check whether SMD force is already present
-        sim.check_for_existing_smd_force()
-
-        if not sim.loaded_smd_force_from_sim:
-            # Create SMD force and add it to the system
-            sim.add_smd_force_to_system()
-
-        sim.simulation.context.reinitialize(preserveState=True)
+        # Initialise all objects relevant to the SMD simulation
+        sim._initialise_smd_simulation(smd_atom_indices, smd_path, smd_force_constant)
 
         # Create a checkpoint of the simulation
         sim.checkpoint = sim.simulation.context.createCheckpoint()
@@ -124,26 +109,12 @@ class OpenMMSMDSimulation:
         # Load the simulation from the path
         with open(sim.xml_path) as infile:
             sim.simulation = serializer.deserialize_simulation(infile)
+
         # Check if simulation employs periodic boundary conditions
[…]
