### 2025-06-27 #8118 @Kevin-Mattheus-Moerman
- [ ] Replace `conda` by `Conda`
