### 2025-06-27 92d82c9
Draft a test for the work done, needs more work...
