### 2025-06-27 195e35c
Add citation suggested by JOSS to the repo
