### 2025-06-27 9e5e126
Update CITATION.cff (#415)
Fix incorrectly formatted name.
