### 2025-06-27 #8118 @hjstroud
Thank you for reviewing our paper @jkrajniak @samuelmurail @abhishektiwari @Kevin-Mattheus-Moerman ! We're also looking forward to supporting JOSS in the future!
