### 2025-06-27 #8118 @hjstroud
@abhishektiwari We've just noticed that one of the names of the authors (Rhoslyn Roebuck Williams) is incorrectly formatted in the .cff file and in the metadata (when you export the citation to Zotero from the paper's page, the name appears incorrectly). Currently, her name is formatted as "Williams, R. R.", but her surname is "Roebuck Williams". Could you change the metadata for the paper to fix this?
