### 2025-06-27 23570e1
Update CITATION.cff (#416)
Put two surnames in quotation marks to ensure that the citation is generated correctly.
