+++ CITATION.cff
@@ -9,7 +9,7 @@ authors:
 - family-names: Barnoud
   given-names: Jonathan
   orcid: "https://orcid.org/0000-0003-0343-7796"
-- family-names: Roebuck Williams
+- family-names: "Roebuck Williams"
   given-names: Rhoslyn
   orcid: "https://orcid.org/0000-0003-2535-7180"
 - family-names: Dhouioui
