### 2025-06-27 df2b515
Update CITATION.cff
Change second instance of Rhos's name
