+++ paper/paper.md
@@ -25,7 +25,7 @@ authors:
     affiliation: '1'
   - name:
       given-names: Rhoslyn 
-      surname: Roebuck Williams
+      surname: "Roebuck Williams"
     orcid: 0000-0003-2535-7180
     affiliation: '1'
   - name: Mohamed Dhouioui
