### 2025-06-27 #8118 @hjstroud
@danielskatz Thank you for clarifying that. I have now pushed a change in the paper.md file to the branch, changing Rhos' name to

`- name:
      given-names: Rhoslyn
      surname: Roebuck Williams`

Hopefully this is sufficient, unless both surnames need to be in quotation marks (i.e. "Roebuck Williams")? All of the other names appear correctly, so they don't require any editing. Sorry about this @Kevin-Mattheus-Moerman !
