### 2025-06-30 9d4a26b
Add JOSS Markdown badge to README.md
