+++ README.md
@@ -1,3 +1,5 @@
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.08118/status.svg)](https://doi.org/10.21105/joss.08118)
+
 # NanoVer Python Server + gRPC Protocol
 
 Repository containing the gRPC protocol and python based implementations
