### 2025-06-30 25de22e
Add badges for tests and docs
