+++ README.md
@@ -1,4 +1,6 @@
 [![DOI](https://joss.theoj.org/papers/10.21105/joss.08118/status.svg)](https://doi.org/10.21105/joss.08118)
+[![Build Status](https://github.com/IRL2/nanover-server-py/actions/workflows/tests.yml/badge.svg?branch=main)](https://github.com/IRL2/nanover-server-py/actions/workflows/tests.yml?query=branch%3Amain)
+[![Docs](https://img.shields.io/badge/docs-latest-blue.svg)](https://irl2.github.io/nanover-docs)
 
 # NanoVer Python Server + gRPC Protocol
 
