### 2025-06-30 951372a
Move badges below title and add badge for license
