### 2025-06-30 8df73d9
Reword sentence before NanoVer citation
