+++ README.md
@@ -171,7 +171,7 @@ If you are having trouble autoconnecting to servers, you can run `nanover-essd-l
 
 ## Citation and external libraries
 
-If you find this project useful, please cite the following paper:
+Any work that uses NanoVer should cite the following publication:
 
 > Stroud, H. J., Wonnacott, M. D., Barnoud, J., Roebuck Williams, R., Dhouioui, M., McSloy, A., Aisa, L., Toledo, L. E., Bates, P., Mulholland, A. J., & Glowacki, D. R. (2025). NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality. Journal of Open Source Software, 10(110), 8118. https://doi.org/10.21105/joss.08118
 
