### 2025-06-30 d55fa4f
Combine references into single suggested citation list
