+++ README.md
@@ -171,11 +171,9 @@ If you are having trouble autoconnecting to servers, you can run `nanover-essd-l
 
 ## Citation and external libraries
 
-Any work that uses NanoVer should cite the following publication:
+If you find this project useful, please cite the following papers:
 
 > Stroud, H. J., Wonnacott, M. D., Barnoud, J., Roebuck Williams, R., Dhouioui, M., McSloy, A., Aisa, L., Toledo, L. E., Bates, P., Mulholland, A. J., & Glowacki, D. R. (2025). NanoVer Server: A Python Package for Serving Real-Time Multi-User Interactive Molecular Dynamics in Virtual Reality. Journal of Open Source Software, 10(110), 8118. https://doi.org/10.21105/joss.08118
-
-As NanoVer builds on the work of Narupa, we would appreciate it if you also cite the following papers:
 
 > Jamieson-Binnie, A. D., O’Connor, M. B., Barnoud, J., Wonnacott, M. D., Bennie, S. J., & Glowacki, D. R. (2020, August 17). Narupa iMD: A VR-Enabled Multiplayer Framework for Streaming Interactive Molecular Simulations. ACM SIGGRAPH 2020 Immersive Pavilion. SIGGRAPH ’20: Special Interest Group on Computer Graphics and Interactive Techniques Conference. https://doi.org/10.1145/3388536.3407891
 
