### 2025-06-30 730a86a
Reword sentence before citations
