### 2025-06-30 f235229
pass on convert_units parameter to mdanalysis trajectory readers
