+++ python-libraries/nanover-server/src/nanover/mdanalysis/universe.py
@@ -91,7 +91,7 @@ KEY_TO_ATTRIBUTE = {
 }
 
 
-def universe_from_recording(*, traj: PathLike[str]):
+def universe_from_recording(*, traj: PathLike[str], convert_units=True):
     """
     Read and convert a NanoVer trajectory recording into an mdanalysis Universe, ignore all frames after a frame_index
     reset.
@@ -100,10 +100,11 @@ def universe_from_recording(*, traj: PathLike[str]):
         traj,
         format=NanoverReader,
         topology_format=NanoverParser,
+        convert_units=convert_units,
     )
 
 
-def universes_from_recording(*, traj: PathLike[str]):
+def universes_from_recording(*, traj: PathLike[str], convert_units=True):
     """
     Decompose a NanoVer trajectory recording into an mdanalysis Universe for each session of simulation (determined
     by frame_index resets).
@@ -125,7 +126,9 @@ def universes_from_recording(*, traj: PathLike[str]):
 
         try:
             universe = frame_data_to_mdanalysis(first_particle_frame)
-            universe.trajectory = NanoverReaderBase(reader, filename=traj)
+            universe.trajectory = NanoverReaderBase(
+                reader, filename=traj, convert_units=convert_units
+            )
             universes.append(universe)
         except Exception as e:
             warnings.warn(f"Failed to extract one universe from recording: {e}")
