### 2025-07-01 6ad887d
Correct the order of operations in `run_smd` to match logic of the numerical approach
