### 2025-07-01 32a808a
Add test to check that the cumulative work done is correctly calculated according to the equation for calculating work done in the SMD procedure
