### 2025-07-01 57b7e5d
Add test to check that the state of a simulation is correctly loaded by the OpenMMSMDSimulation class and format with Black
