### 2025-07-01 e397cfe
Add comments and a relative tolerance to test that tests state loading
