### 2025-07-01 cbe63e2
Add new test to list of things to test
