### 2025-07-01 094ef1b
Rename OpenMM SMD test file
