### 2025-07-01 d1da841
Update "test_load_openmm_state" to test `save_simulation()` too
