### 2025-07-01 8c2afa6
Redirect output in `test_load_openmm_state`
