### 2025-07-01 4ea0067
Replace atom guessing in `PathSmoother` to remove depreciation warning
