### 2025-07-05 #8118 @Kevin-Mattheus-Moerman
@openjournals/dev @danielskatz can I close this issue now, or do I need to reaccept?
