### 2025-07-09 55a5615
Fix units printed in function and try increasing fidelity of smoothing parameter
