### 2025-07-10 90e2215
fix bug with state splitting, include simulation name if it exists
